% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc4Data.R
\docType{data}
\name{allFeatureBasedData}
\alias{allFeatureBasedData}
\title{allFeatureBasedData}
\format{A data frame with 497680 rows and 3 variables}
\usage{
data(allFeatureBasedData)
}
\value{
A data set with 497680 rows and 3 variables
}
\description{
A data set from gene feature based differential analysis
}
\details{
This gene feature based table includes the following information:

\itemize{
  \item geneID: Gene ID
  \item countbinID: ID for features of each gene(it could be exons or splicing junctions)
  \item pvalue: p-value of the features from differential usage analysis 
}
}
\keyword{datasets}
