% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callQoRT.R
\name{callQoRT}
\alias{callQoRT}
\title{{callQoRT}}
\usage{
callQoRT(input_file, runing_cmd, gtf_file)
}
\arguments{
\item{{cmd1}}{}

\item{{input}{bam file}}

\item{{gene}{annotation file}}
}
\value{
a list include counts for each sample
}
\description{
{Use QoRTs to get the count of subfeatures in each gene}
}
\examples{

#Input files
dir.name="/media/H_driver/Aimin_project/GOSJ_STAR_Bam/"

dir.name="/media/H_driver/Aimin_project/GOSJ_STAR_Bam/"
file.name=dir(dir.name,recursive = TRUE,pattern="sorted.bam")
file.name.whole<-paste0(dir.name,file.name)
file.name.selected<-file.name.whole
file.name.selected.2<-as.list(file.name.selected)
names(file.name.selected.2)=sapply(strsplit(file.name.selected,split="\\\\/"),"[[",6)
file.name.selected.3<-file.name.selected.2

cmd2="java -Xmx5000M -jar /home/aiminyan/QoRTs/QoRTsFullExampleData/QoRTsRelease/QoRTs.jar QC --noGzipOutput"
gtf1="/media/H_driver/Aimin_project/GTF_Files/Homo_sapiens.GRCh38.84.processed.sorted.2.gtf"

re.out<-lapply(file.name.selected.3,callQoRT,gtf_file=gtf1,runing_cmd=cmd2)

}

