% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckBias.R
\name{Cbs}
\alias{Cbs}
\title{Cbs}
\usage{
Cbs(re.gene.based, ad = "GL", sub_feature = NULL, threshold, genomeID,
  geneID, gene_model, method)
}
\arguments{
\item{re.gene.based:}{Gene based table}

\item{ad:}{The possible bias factor}

\item{sub_feature:}{The possible bias factor}

\item{threshold:}{threshold used}

\item{genomeID:}{which genome}

\item{geneID:}{which type of gene ID}

\item{gene_model:}{Gene model}

\item{method:}{which method}
}
\description{
Cbs is to check possible bias factor using the method in goseq
}
\examples{

#library(PathwaySplice)
#data(mds)
#Cbs(mds,ad="E",sub_feature="E",threshold=0.05)


}

