% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc4Data.R
\docType{data}
\name{chr22}
\alias{chr22}
\title{chr22 runJunctionSeqAnalyses data set}
\format{A object of class \code{"runJunctionSeqAnalyses"}}
\usage{
data(chr22)
}
\description{
a data set for chr22
}
\keyword{datasets}

