% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combine3Re.R
\name{Combine3Re}
\alias{Combine3Re}
\title{Combine3Re}
\usage{
Combine3Re(Re.PJ, re.PJ.gene.based, re.rMAT, gene.model, Output_files_dir)
}
\arguments{
\item{Re.PJ:}{Output from JunctionSeq}

\item{re.PJ.gene.based:}{Gene based results from JunctionSeq}

\item{re.rMAT:}{Results from rMAT}

\item{Output_GO_BP_feature_gene_redefined}{}
}
\description{
This function combines feature based differential usage results, the derived gene based results from feature based results and
the resutls from rMAT to identify the common results between 3 methods
}
\examples{
Re.combine.3.methods<-Combine3Re(Re.PJ, re.PJ.gene.based, re.rMAT,gene.model,output.dir.name.new)

Re.combine.3.methods.rMAT.SE<-Combine3Re(Re.PJ, re.PJ.gene.based, re.rMAT,gene.model,output.dir.name.rMAT_SE_based)


}

