% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineCpTftGoGeneID.R
\name{CombineCpTftGoGeneID}
\alias{CombineCpTftGoGeneID}
\title{CombineCpTftGoGeneID}
\usage{
CombineCpTftGoGeneID(cp_input, tft_input, Go_input, gene_anno)
}
\description{
CombineCpTftGoGeneID
}
\examples{

gene.id.all<-CombineCpTftGoGeneID("/media/H_driver/Annotation/MsigDB/c2.cp.Mouse.v5.1.symbols.gmt",
"/media/H_driver/Annotation/MsigDB/c3.tft.Mouse.v5.1.symbols.gmt","/media/H_driver/Annotation/GO_gene_mouse/gene_GO.mgi",
"/media/H_driver/Annotation/mm10/genes_table_02052016.csv")


}

