% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareDEfromGeneWithDEfromFeatures.R
\name{CompareDEfromGeneWithDEfromFeatures}
\alias{CompareDEfromGeneWithDEfromFeatures}
\title{CompareDEfromGeneWithDEfromFeatures}
\usage{
CompareDEfromGeneWithDEfromFeatures(Re.PJ.gene, re.PJ.gene.based,
  Output_venn_file)
}
\arguments{
\item{Re.PJ.gene:}{gene-based DE results}

\item{re.PJ.gene.based:}{feature-based DE results}

\item{Output_venn_file:}{file name for venn}
}
\description{
This function compares gene-based DE with feature-based DE
}
\examples{

Gene.based.DE.feature.based.DE<-CompareDEfromGeneWithDEfromFeatures(Re.PJ.gene,re.PJ.gene.based,"/media/H_driver/PJ/gene_feature_DE.tiff")

}

