% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareGOResults.R
\name{CompareGOResults}
\alias{CompareGOResults}
\title{CompareGOResults}
\usage{
CompareGOResults(Gene.based.DE.feature.based.DE, gene_model)
}
\arguments{
\item{gene_model}{}
}
\description{
CompareGOResults
}
\examples{


}

