% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draw4Cat.R
\name{Draw4Cat}
\alias{Draw4Cat}
\title{Draw4Cat}
\usage{
Draw4Cat(RE.exon.sj.all.gene.enriched)
}
\arguments{
\item{RE.exon.sj.all.gene}{}
}
\description{
Draw4Cat
}
\examples{
Draw4Cat(RE.exon.sj.all.gene)

}

