% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawFeature.R
\name{DrawFeature}
\alias{DrawFeature}
\title{DrawFeature}
\usage{
DrawFeature(Re.PJ, out.dir, gene_id)
}
\arguments{
\item{gene_id}{}
}
\description{
DrawFeature
}
\examples{

DrawFeature(Re.PJ,output.dir.name,"ENSMUSG00000026563+ENSMUSG00000040596+ENSMUSG00000089853+ENSMUSG00000103400")


}

