% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichMap_test.R
\name{enMap}
\alias{enMap}
\title{enMap}
\usage{
enMap(x, n = 50, fixed = TRUE, vertex.label.font = 1, SimilarityThreshold,
  ...)
}
\arguments{
\item{x}{gseaResult or enrichResult object}

\item{n}{maximum number of category to shown}

\item{fixed}{if set to FALSE, will invoke tkplot}

\item{vertex.label.font}{font size of vertex label}

\item{SimilarityThreshold}{threshold for defining similarity between GOs}

\item{...}{additional parameter}
}
\value{
figure
}
\description{
enMap is used to draw network based on similarities between GOs
}
\author{
Aimin created this funciton based on enrichMap function in G Yu's DOSE R package
}

