% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichMap_test.R
\name{enrichMap.test}
\alias{enrichMap.test}
\title{enrichMap}
\usage{
enrichMap.test(x, n = 50, fixed = TRUE, vertex.label.font = 1, ...)
}
\arguments{
\item{x}{gseaResult or enrichResult object}

\item{n}{maximum number of category to shown}

\item{fixed}{if set to FALSE, will invoke tkplot}

\item{vertex.label.font}{font size of vertex label}

\item{...}{additional parameter}
}
\value{
figure
}
\description{
enrichment map
}
\details{
enrichment map
}
\author{
Aimin did some modificatios based G Yu's code in DOSE
}

