% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{enrichmentMap}
\alias{enrichmentMap}
\title{enrichmentMap}
\usage{
enrichmentMap(goseqres, n = 50, fixed = TRUE, vertex.label.font = 1,
  similarity.threshold, output.file.dir, label.vertex.by.index = FALSE, ...)
}
\arguments{
\item{goseqres}{Object returned from runPathwaySplice}

\item{n}{The top \emph{n} categories are shown in enrichment map}

\item{fixed}{If set to FALSE, will invoke tkplot}

\item{vertex.label.font}{Font size of vertex label}

\item{similarity.threshold}{Gene sets with Jaccard Coefficient > similarity.threshold will be connected on the enrichment map}

\item{output.file.dir}{Output dir for the gene set information file on network}

\item{label.vertex.by.index}{Which way to be used for labeling vertex on network

FALSE indicates to label vertex by the name of gene sets

TRUE indicates to label vertex by the index of gene sets}

\item{...}{Additional parameter}
}
\value{
A list for giving edge and vertex information of enrichment map
}
\description{
This function draws an enrichment map based on similarities 
between gene sets based on Jaccard similarity Coefficient
}
\details{
The Jaccard similarity coefficient ranges from 0 to 1. JC=0 indicates 
there are no overlapping genes between two gene sets, 
JC=1 indicates two gene sets are identical.
}
\examples{

gene.based.table <- makeGeneTable(featureBasedData)

res <- runPathwaySplice(gene.based.table,genome='hg19',
                         id='ensGene',test.cats=c('GO:BP'),
                         go.size.limit=c(5,30),
                         method='Wallenius')

output.file.dir <- file.path(tempdir(),'OutputEnmapEx')

enmap <- enrichmentMap(res,n=10,similarity.threshold=0,
                       output.file.dir = output.file.dir,
                       label.vertex.by.index = TRUE)
                       
}
\author{
Aimin created this funciton based on enrichMap function in G Yu's DOSE R package
}
