% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractGeneIDFromGo.R
\name{ExtractGeneIDFromGo}
\alias{ExtractGeneIDFromGo}
\title{ExtractGeneIDFromGo}
\usage{
ExtractGeneIDFromGo(gene_go_file, gene_anno_file)
}
\description{
ExtractGeneIDFromGo
}
\examples{
gene.id.from.go<-ExtractGeneIDFromGo("/media/H_driver/Annotation/GO_gene_mouse/gene_GO.mgi","/media/H_driver/Annotation/mm10/genes_table_02052016.csv")


}

