% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc4Data.R
\docType{data}
\name{gene.model.hg38}
\alias{gene.model.hg38}
\title{hg38 Gene annotation data set}
\format{A data frame with 24323 rows and 6 variables}
\usage{
data(gene.model.hg38)
}
\value{
a data set 24323 rows and 6 variables
}
\description{
a data set for hg38 gene annotation
}
\keyword{datasets}

