% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene2cat.R
\name{gene2cat}
\alias{gene2cat}
\title{Given a gene name and the object built from reading Gmt file, and find the pathways that this gene corresponds to}
\usage{
gene2cat(gene_name, re)
}
\arguments{
\item{re}{}
}
\description{
Given a gene name and the object built from reading Gmt file, and find the pathways that this gene corresponds to
}

