% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene2cat2.R
\name{gene2cat2}
\alias{gene2cat2}
\title{Read gmt file, and convert this gmt file to a list with the name of each element being gene name
and each element being the pathways that this gene corresponds to}
\usage{
gene2cat2(gmt_input_file)
}
\arguments{
\item{gmt_input_file}{}
}
\description{
Read gmt file, and convert this gmt file to a list with the name of each element being gene name
and each element being the pathways that this gene corresponds to
}

