% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneBasedAnalysis.R
\name{GeneBasedAnalysis}
\alias{GeneBasedAnalysis}
\title{GeneBasedAnalysis}
\usage{
GeneBasedAnalysis(Re.PJ.gene, gene_model, output_file)
}
\arguments{
\item{Re.PJ.gene:}{gene-based DE results}

\item{gene_model:}{gene model}

\item{output_file:}{file name for venn}
}
\description{
This function use gene-based DE to get GO
}
\examples{

Gene.non.coding.GO<-GeneBasedAnalysis(Re.PJ.gene.rm.non.coding,gene.model,"/media/H_driver/PJ/geneGL_rm_non_coding.xls")


}

