% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateGeneAnno.R
\name{GenerateGeneAnno}
\alias{GenerateGeneAnno}
\title{Generate gene annotations based on the Ensembl ID of genes}
\usage{
GenerateGeneAnno(TobeAnno, data.set)
}
\arguments{
\item{TobeAnno:}{gene list}
}
\description{
Generate gene annotations based on the Ensembl ID of genes
}
\examples{

Cheng.gene.all.anno.3<-do.call(rbind,lapply(re.PJ.gene.based.testable[,1],GenerateGeneAnno,gene.model))


}

