% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneWisePvsExonWiseP.R
\name{GeneWisePvsExonWiseP}
\alias{GeneWisePvsExonWiseP}
\title{Compare gene based p value with exon based p value}
\usage{
GeneWisePvsExonWiseP(input_data_set, gene_wise_p_c, feature_wise_p_c)
}
\arguments{
\item{input_data_set}{data set for all genes}

\item{gene_wise_p_c}{genewise p value column}

\item{feature_wise_p_c}{feature(exon or SJ) p value column}
}
\examples{

load("/media/H_driver/2015/Nimer_Cheng/Data_set_two_methods.RData")
GeneWisePvsExonWiseP(data.table.gene.based.all.jscs4,7,9)

}

