% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Over_under_represented_pvalue.R
\name{Get_Over_under_represented_pvalue}
\alias{Get_Over_under_represented_pvalue}
\title{Title}
\usage{
Get_Over_under_represented_pvalue(num_de_incat, num_incat, num_genes, num_de,
  weight)
}
\arguments{
\item{weight}{}
}
\description{
Title
}
\examples{

28 	 186 	 1852 	 125 	 1.03043

Get_Over_under_represented_pvalue(28,186,1852,125,1.03043)

Get_Over_under_represented_pvalue(28,186,1840,125,1.080766)

}

