% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdaptedCodes_juncbias.R
\name{getDataset}
\alias{getDataset}
\title{codes for simulation and plot PWF
simulation code from xiaobei
PWF plot code from goseq
Get data set for simulation studies}
\usage{
getDataset(counts, drop.extreme.dispersion = 0.1, drop.low.lambda = TRUE)
}
\arguments{
\item{drop.low.lambda}{}
}
\description{
codes for simulation and plot PWF
simulation code from xiaobei
PWF plot code from goseq
Get data set for simulation studies
}

