% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDesign.R
\name{GetDesign}
\alias{GetDesign}
\title{Title}
\usage{
GetDesign(sample.names, condition, use.covars = NULL)
}
\arguments{
\item{use.covars}{}
}
\description{
Title
}
\examples{

dir.name="/Volumes/Bioinformatics$/2015/Nimer_Cheng/"
file.sample="decoder.bySample.txt"
path.file.sample<-paste0(dir.name,file.sample)
decoder.bySample<-read.table(path.file.sample,header=T,stringsAsFactors = F)
print(decoder.bySample)
Re<-GetDesign(decoder.bySample$sample.ID,decoder.bySample$group.ID)

}

