% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_function_4_PathwaySplice.R
\name{getGeneSet}
\alias{getGeneSet}
\title{gene.4.go <- res2$geneID
go <- getGeneSet(gene.4.go,"hg19","ensGene",fetch.cats = c("GO:CC"),go.size.cut=c(5,10))}
\usage{
getGeneSet(genes, genome, id, fetch.cats = c("GO:CC", "GO:BP", "GO:MF"),
  go.size.cut)
}
\description{
gene.4.go <- res2$geneID
go <- getGeneSet(gene.4.go,"hg19","ensGene",fetch.cats = c("GO:CC"),go.size.cut=c(5,10))
}
