% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_function_4_PathwaySplice.R
\name{getGeneSetBySize}
\alias{getGeneSetBySize}
\title{getGeneSetBySize}
\usage{
getGeneSetBySize(user2cat, go.size.cut)
}
\arguments{
\item{user2cat}{A list with its name being gene name , and its element being gene sets that gene belongs to}

\item{go.size.cut}{Size of gene set users can define}
}
\description{
getGeneSetBySize
}
\examples{

}
