% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMgiSymbolUsingLocalDataBase.R
\name{GetMgiSymbolUsingLocalDataBase}
\alias{GetMgiSymbolUsingLocalDataBase}
\title{Title}
\usage{
GetMgiSymbolUsingLocalDataBase(data.gene, data.set)
}
\arguments{
\item{data.gene}{}
}
\description{
Title
}
\examples{


}

