% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPwfUseReformatedData.R
\name{GetPwfUseReformatedData}
\alias{GetPwfUseReformatedData}
\title{GotermAnalysisUseReformatedData}
\usage{
GetPwfUseReformatedData(re.gene.based, ad = "GL", sub_feature = NULL,
  threshold)
}
\arguments{
\item{threshold}{}
}
\description{
GotermAnalysisUseReformatedData
}
\examples{

Re.pwf.exon.sj<-GetPwfUseReformatedData(ReformatData(re.PJ.gene.based),ad="exon_SJ",sub_feature=NULL,0.05)



}

