% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetResults4GeneLevel.R
\name{GetResults4GeneLevel}
\alias{GetResults4GeneLevel}
\title{GetResults4GeneLevel}
\usage{
GetResults4GeneLevel(dir.name, file.sample, file.count)
}
\arguments{
\item{dir.name:}{path that count file and sample information file is in}

\item{file.sample:}{sample information file}

\item{file.count:}{count files}
}
\value{
results from DESeq2
}
\description{
This function applies DESeq2 to gene level count to identify differentially expressed genes
}
\examples{
# For PJ project
dir.name.PJ.gene="/media/H_driver/PJ/"
file.sample.PJ.gene="decoder.bySample.rtf"
file.count.PJ.gene="/QC.geneCounts.formatted.for.DESeq.txt"

Re.PJ.gene<-GetResults4GeneLevel(dir.name.PJ.gene,file.sample.PJ.gene,file.count.PJ.gene)
}

