% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_get_results_from_JunctionSeq.R
\name{GetResultsFromJunctionSeq}
\alias{GetResultsFromJunctionSeq}
\title{Get analysis results using JunctionSeq}
\usage{
GetResultsFromJunctionSeq(dir.name, file.sample, file.count, file.gff)
}
\arguments{
\item{file.gff}{}
}
\description{
Get analysis results using JunctionSeq
}

