% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goseq2.R
\name{goseq2}
\alias{goseq2}
\title{goseq2}
\usage{
goseq2(pwf, genome, id, gene.model, gene2cat = NULL, test.cats = c("GO:CC",
  "GO:BP", "GO:MF"), method = "Wallenius", repcnt = 2000,
  use_genes_without_cat = FALSE, outputfile)
}
\arguments{
\item{pwf:}{probability weight function}

\item{genome:}{genome you use}

\item{id:}{which gene id}

\item{gene2cat:}{a list with gene as name and category as value}

\item{test.cats:}{the category you choose}

\item{method:}{which method for calculting overpresentation p-value}

\item{repcnt:}{the number of replications}

\item{use_genes_without_cat:}{whether using genes without category or not}
}
\description{
Modifying goseq to generate GO term with the listed DEgene
}
\examples{
test.goseq2<-goseq2(Re.Go.adjusted.by.number.junction.2[[2]],"mm10","ensGene",gene.model)

test.goseq33<-goseq2(Re.Go.adjusted.by.number.junction.2[[2]],"mm10","ensGene",gene.model,
outputfile="/media/H_driver/PJ/GO_term_adjusted_by_SJ_2.xls")

test.goseq44<-goseq2(Re.Go.adjusted.by.number.junction.2[[2]],"mm10","ensGene",gene.model,
outputfile="GO_term_adjusted_by_SJ_3.xls")

sink("output_dhyper.txt")
test.goseq44<-goseq2(Re.Go.adjusted.by.number.junction.2[[2]],"mm10","ensGene",gene.model,
outputfile="GO_term_adjusted_by_SJ_4.xls")
sink()

sink("output_dhyper_use_all.txt")
test.goseq55<-goseq2(Re.Go.adjusted.by.number.junction.2[[2]],"mm10","ensGene",gene.model,
outputfile="GO_term_adjusted_by_SJ_use_all.xls")
sink()


}

