% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotermAnalysis2GetAllGOTerms_AdjustedByNumOfJunctionWithinOneGene.R
\name{GotermAnalysis2GetAllGOTerms_AdjustedByNumOfJunctionWithinOneGene}
\alias{GotermAnalysis2GetAllGOTerms_AdjustedByNumOfJunctionWithinOneGene}
\title{This function is to subset data set based on feature}
\usage{

  GotermAnalysis2GetAllGOTerms_AdjustedByNumOfJunctionWithinOneGene(re.gene.based,
  ad = "GL", sub_feature = NULL, threshold, gene_model)
}
\arguments{
\item{threshold}{}
}
\description{
This function is to subset data set based on feature
}
\examples{


Re.Go.adjusted.by.number.junction.2<-GotermAnalysis2GetAllGOTerms_AdjustedByNumOfJunctionWithinOneGene(
re.PJ.gene.based,ad="J","J",0.05,"Splice_junction_based")

data.pwf2.SJs<-plotPWF2(Re.Go.adjusted.by.number.junction.2[[2]],binsize=30,xlab = "Number of SJs(<binsize> gene bins)")

Re.Go.adjusted.by.exon.SJ<-GotermAnalysis2GetAllGOTerms_AdjustedByNumOfJunctionWithinOneGene(
re.PJ.gene.based,ad="exon_SJ",sub_feature=NULL,0.05,file_prefix="Exon_Splice_junction_based.xls",gene_model=gene.model)

Re.Go.adjusted.by.exon.SJ<-GotermAnalysis2GetAllGOTerms_AdjustedByNumOfJunctionWithinOneGene(
mds,ad="exon_SJ",sub_feature="E",0.05,gene_model="hg19")

}

