% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotermAnalysisUseReformatedData.R
\name{GotermAnalysisUseReformatedData}
\alias{GotermAnalysisUseReformatedData}
\title{GotermAnalysisUseReformatedData}
\usage{
GotermAnalysisUseReformatedData(re.gene.based, ad = "GL",
  sub_feature = NULL, threshold, genomeID, geneID, gene_model)
}
\arguments{
\item{threshold}{}
}
\description{
GotermAnalysisUseReformatedData
}
\examples{
data(mds)
data(hg19)
Example.Go.adjusted.by.exon<-GotermAnalysisUseReformatedData(mds,ad="exon_SJ",sub_feature="E",0.05,genomeID="hg19",geneID="ensGene",gene_model=hg19.gene.model)

}

