% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc4Data.R
\docType{data}
\name{hg19.gene.model}
\alias{hg19.gene.model}
\title{hg19 Gene annotation data set}
\format{A data frame with 64437 rows and 6 variables}
\usage{
data(hg19)
}
\value{
a data frame with 64437 rows and 6 variables
}
\description{
a data set for hg19 gene annotation
}
\keyword{datasets}

