% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc4Data.R
\docType{data}
\name{hg19}
\alias{hg19}
\title{hg19 Gene annotation data set}
\format{A data frame with 64437 rows and 6 variables}
\usage{
data(hg19)
}
\value{
a data frame with 64437 rows and 6 variables
}
\description{
a data set for hg19 gene annotation
  
\itemize{
   \item hgnc_symbol. hgnc gene symbol  
   \item entrezgene. Entrez gene ID
   \item ensembl_gene_id. Ensembl gene ID
   \item chromosome_name. Chromosome information
   \item start_position. Starting position
   \item end_position. End position
}
}
\keyword{datasets}
