% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc4Data.R
\docType{data}
\name{hg38}
\alias{hg38}
\title{hg38 Gene annotation data set}
\format{A data frame with 24323 rows and 6 variables}
\usage{
data(hg38)
}
\value{
a data set 24323 rows and 6 variables
}
\description{
a data set for hg38 gene annotation
}
\details{
\itemize{
   \item gene_id. Gene symbol  
   \item entrezgene. Entrez gene ID
   \item ensembl_gene_id. Ensembl gene ID
   \item chromosome_name. Chromosome information
   \item start_position. Starting position
   \item end_position. End position
}
}
\keyword{datasets}
