% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JS.perGeneQValue.R
\name{JS.perGeneQValue}
\alias{JS.perGeneQValue}
\title{Title}
\usage{
JS.perGeneQValue(pvals, wTest, geneID, method = JS.perGeneQValueExact)
}
\arguments{
\item{method}{}
}
\description{
Title
}
\examples{

gene.1<-results.with.count.data.2[which(results.with.count.data.2[,1]=="ENSMUSG00000000028+ENSMUSG00000005262"),]

results.with.count.data.2[which(results.with.count.data.2[,1]=="ENSMUSG00000000028+ENSMUSG00000005262"),]

sink("output.txt")
data.perGene<-JS.perGeneQValue(results.with.count.data.2$pvalue,results.with.count.data.2$testable,results.with.count.data.2$geneID,method = JS.perGeneQValueExact)
sink()
length(data.perGene)


data.perGene.1<-JS.perGeneQValue(gene.1$pvalue,gene.1$testable,gene.1$geneID,method = JS.perGeneQValueExact)
length(data.perGene.1)

}

