% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JS.perGeneQValueExact.R
\name{JS.perGeneQValueExact}
\alias{JS.perGeneQValueExact}
\title{JS.perGeneQValueExact}
\usage{
JS.perGeneQValueExact(pGene, theta, geneSplit)
}
\arguments{
\item{pGene:p}{value}

\item{geneSplit:split}{}
}
\description{
Exact computation - see methods part of the paper
}
\examples{


}

