% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JS.perGeneQValueExact_test.R
\name{JS.perGeneQValueExact.test}
\alias{JS.perGeneQValueExact.test}
\title{Title}
\usage{
JS.perGeneQValueExact.test(pGene, theta, geneSplit)
}
\arguments{
\item{geneSplit}{}
}
\description{
Title
}
\examples{
source("https://bioconductor.org/biocLite.R")
biocLite("Biobase")
library(Biobase)

}

