% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabelGeneBasedFeature.R
\name{LabelGeneBasedFeature}
\alias{LabelGeneBasedFeature}
\title{LabelGeneBasedFeature}
\usage{
LabelGeneBasedFeature(Re.PJ, cutoff_FC, cutoff_p, Output_file)
}
\arguments{
\item{Re.PJ}{}
}
\description{
LabelGeneBasedFeature
}
\examples{
Re.PJ.selected<-SelectFeature(Re.PJ)

dim(Re.PJ.selected)

LabelGeneBasedFeature(Re.PJ,cutoff_FC,cutoff_p,Output_file)


}

