% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lcb.R
\name{Lcb}
\alias{Lcb}
\title{LrCheckBias}
\usage{
Lcb(jscs_genewise_object, genewise.pvalue = "geneWisePadj",
  sig.threshold = 0.05, type = c("exon", "splicing"))
}
\arguments{
\item{genewise.pvalue:}{whether you use smallest p value or geneWisePadj}

\item{sig.threshold:}{threshold to define differential gene list}

\item{type:}{whether you are interested in exon or splicing junction}

\item{jscs_genewise_object:}{an object returned from ReformatData and MakeGeneWiseTable_JunctionSeq}
}
\value{
return results from logistic regression
}
\description{
Logistic regression to check bias
}
\examples{

data(mds)
re<-Lcb(mds)

}

