% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{lrTestBias}
\alias{lrTestBias}
\title{lrTestBias}
\usage{
lrTestBias(jscs.genewise.object, genewise.pvalue = "geneWisePadj",
  sig.threshold = 0.05, type = c("exon", "splicing"), loc.x = 2,
  loc.y = 70, y.lim = 80, boxplot.width)
}
\arguments{
\item{jscs.genewise.object}{Gene based table}

\item{genewise.pvalue}{Whether you use smallest p value or geneWisePadj}

\item{sig.threshold}{Threshold to define differential gene list}

\item{type}{Whether you are interested in exon or splicing junction}

\item{loc.x}{Indicates x coordinate of p value from logistic regression}

\item{loc.y}{Indicates y coordinate of p value from logistic regression}

\item{y.lim}{Defining the largest number of exons in y axis in boxplot}

\item{boxplot.width}{Parameter for boxplot width}
}
\value{
Results from logistic regression
}
\description{
Logistic regression to check bias
}
\examples{
res <- lrTestBias(tiny.data,loc.x=2,loc.y=150,y.lim=200,boxplot.width=0.3)
}
