% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{makeFeatureTable}
\alias{makeFeatureTable}
\title{makeFeatureTable}
\usage{
makeFeatureTable(jscs)
}
\arguments{
\item{jscs}{An \code{JunctionSeqCountSet} object returned from function \code{getResultsFromJunctionSeq},
with restuls generated by testing differential usage of gene features (i.e. exon counting bins and/or splice junction counting bins)}
}
\value{
A dataframe with the following arguments as cloumns for each feature 
  \itemize{
  \item geneID: Gene ID
  \item countbinID: feature ID in the feature differential usage analysis
  \item pvalue: unadjusted p value of each feature in the feature differential usage analysis
}
}
\description{
This function generates a dataframe that summarizes differential usage of features within the gene.
}
\examples{

#First test differential usage for exons or splicing junctions
dir.name <- system.file('extdata',package = 'PathwaySplice')
sample.file <- 'Sample_info.txt'
count.file <- 'Counts.10.genes.txt'
gff.file <- 'flat.chr22.10.genes.gff'
res <- getResultsFromJunctionSeq(dir.name, sample.file, 
                                 count.file, gff.file, 
                                 method.dispFinal='max')
                                 
# Convert the results of differential usage analysis into a gene table
res1 <- PathwaySplice:::makeFeatureTable(res)

}
