% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{makeGeneTable}
\alias{makeGeneTable}
\title{res2 <- PathwaySplice:::makeGeneTable(res1)}
\usage{
makeGeneTable(feature.table, sig.threshold = 0.05)
}
\description{
res2 <- PathwaySplice:::makeGeneTable(res1)
}
