% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeGeneWiseTable_JunctionSeq.R
\name{makeGeneWiseTable}
\alias{makeGeneWiseTable}
\title{This function is to generate the gene-based results}
\usage{
makeGeneWiseTable(jscs, gene.list, FDR.threshold = 0.05, verbose = TRUE,
  debug.mode = FALSE)
}
\arguments{
\item{debug.mode}{}
}
\description{
This function is to generate the gene-based results
}
\examples{

load("Re_Run_test_GOSJ.RData")

re.gene.based<-makeGeneWiseTable(Re,gene.list=unique(as.character(fData(Re)$geneID)))
dim(pData(re.gene.based))
no.testable.index<-which(as.character(pData(re.gene.based)$mostSigID)=="character(0)")
re.gene.based.testable<-pData(re.gene.based)[-no.testable.index,]
dim(re.gene.based.testable)

cor(as.numeric(re.gene.based.testable$numKnown),
as.numeric(re.gene.based.testable$mostSigPadjust))

}

