% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc4Data.R
\docType{data}
\name{mds.11.sample}
\alias{mds.11.sample}
\title{A gene-based result data set by using 11 subjuects(excluding 309)}
\format{A data frame with 23520 rows and 17 variables}
\usage{
data(mds11)
}
\value{
a data set 23520 rows and 17 variables
}
\description{
a gene-based result data set
}
\details{
\itemize{
  \item geneID. Gene ID
  \item chr. Chromosome information
  \item start. Starting position
  \item end. End position
  \item strand. Strand information
  \item baseMean. BaseMean based on all samples
  \item geneWisePadj. perGeneQvalue from DEXSeq package 
  \item mostSigID. The feature that have smallest p-value within gene  
  \item mostSigPadjust. p-value for the most significant feature 
  \item numExons. Number of exons within gene
  \item numKnown. Number of splicing junction within gene
  \item numNovel. Number of novel splicing junction within gene
  \item exonsSig. Number of significant exons
  \item knownSig. Number of significant splicing junction
  \item novelSig. Number of significant novel splicing junction
  \item numFeatures. A list that includes number of exons,number of 
  splicing junction and number of novel splicing junction within gene 
  \item numSig. A list that includes number of significant exons,number of 
  significant splicing junction and number of significant novel splicing 
  junction within gene 
}
}
\keyword{datasets}
