% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc4Data.R
\docType{data}
\name{mm10}
\alias{mm10}
\title{mm10 Gene annotation data set}
\format{A data frame with 23440 rows and 6 variables}
\usage{
data(mm10)
}
\value{
a data set 23440 rows and 6 variables
}
\description{
a data set for mm10 gene annotation
 
\itemize{
   \item gene_id. Gene symbol  
   \item entrezgene. Entrez gene ID
   \item ensembl_gene_id. Ensembl gene ID
   \item chromosome_name. Chromosome information
   \item start_position. Starting position
   \item end_position. End position
}
}
\keyword{datasets}
