% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBsim.R
\name{NBsim}
\alias{NBsim}
\title{Generate simulated count based on NB distribution}
\usage{
NBsim(dataset, group, nTags = 10000, nlibs = length(group),
  fix.dispersion = NA, lib.size = NULL, drop.low.lambda = TRUE,
  drop.extreme.dispersion = 0.1, add.outlier = FALSE, outlierMech = c("S",
  "R", "M"), pOutlier = 0.1, min.factor = 1.5, max.factor = 10,
  pDiff = 0.1, pUp = 0.5, foldDiff = 3, name = NULL,
  save.file = FALSE, file = NULL, only.add.outlier = FALSE,
  verbose = TRUE)
}
\arguments{
\item{verbose}{}
}
\description{
Generate simulated count based on NB distribution
}

