% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputGO.R
\name{OuputGO}
\alias{OuputGO}
\title{OutputGO}
\usage{
OuputGO(Re.Go.adjusted.by.exon.SJ, Output_file)
}
\arguments{
\item{Output_file}{}
}
\description{
OutputGO
}
\examples{

Output.GO<-OuputGO(Example.Go.adjusted.by.exon[[1]]$GO,Output_file=paste0(getwd(),"/example.xls"))

Draw4GO(Output.GO)

}

