% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputCatBasedPwf.R
\name{OutputCatBasedPwf}
\alias{OutputCatBasedPwf}
\title{OutputCatBasedPwf}
\usage{
OutputCatBasedPwf(Re.pwf.exon.sj, gene_model, gene_2_cat, Output_file)
}
\arguments{
\item{Output_file}{}
}
\description{
Reformat genewised results from JunctionSeq, and use this one to get enriched GO terms
}
\examples{

RE.cp<-OutputCatBasedPwf(Re.pwf.exon.sj,gene_model=gene.model,gene_2_cat=gene.2.cat.cp.mouse,Output_file="Cp.xls")

RE.cft<-OutputCatBasedPwf(Re.pwf.exon.sj,gene_model=gene.model,gene_2_cat=gene.2.cat.tft.mouse,Output_file="C_tft.xls")


RE.cp.gene<-OutputCatBasedPwf(Gene.non.coding.GO$pwf,gene_model=gene.model,gene_2_cat=gene.2.cat.cp.mouse,
Output_file="/media/H_driver/PJ/geneGL_rm_non_coding_Cp.xls")

RE.cft.gene<-OutputCatBasedPwf(Gene.non.coding.GO$pwf,gene_model=gene.model,gene_2_cat=gene.2.cat.tft.mouse,
Output_file="/media/H_driver/PJ/geneGL_rm_non_coding_tft.xls")


}

