% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputFigure.R
\name{OutputFigure}
\alias{OutputFigure}
\title{Title}
\usage{
OutputFigure(jscs.object, gene.name, outfile_prefix)
}
\arguments{
\item{outfile_prefix}{}
}
\description{
Title
}
\examples{
OutputFigure(jscs.object=Re.PJ,gene.name="ENSMUSG00000000028+ENSMUSG00000005262",
outfile_prefix="/media/H_driver/PJ/plot_ENSMUSG00000000028_ENSMUSG00000005262/")

}

