% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputGeneWiseTable.R
\name{OutputGeneWiseTable}
\alias{OutputGeneWiseTable}
\title{Title}
\usage{
OutputGeneWiseTable(re.PJ.gene.based, gene.model, output_file)
}
\arguments{
\item{output_file}{}
}
\description{
Title
}
\examples{

data(gene.model)

OutputGeneWiseTable(re.PJ.gene.based,gene.model,output_file="/media/H_driver/PJ/GeneWise_Re_annotated_using_new_annotation_3.csv")


}

