% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputGOBasedDEfromFeatures.R
\name{OutputGOBasedDEfromFeatures}
\alias{OutputGOBasedDEfromFeatures}
\title{OutputGOBasedDEfromFeatures}
\usage{
OutputGOBasedDEfromFeatures(re.PJ.gene.based, gene.model, Output_file)
}
\arguments{
\item{Output_file}{}
}
\description{
Use genewised results from JunctionSeq to get enriched GO terms
}
\examples{


RE.exon.sj<-OutputGOBasedDEfromFeatures(re.PJ.gene.based,gene.model,"GO_exon_sj_3.xls")

sink("test.txt")
RE.exon.sj<-OutputGOBasedDEfromFeatures(re.PJ.gene.based,gene.model,"GO_exon_sj_4.xls")
sink()

}

