% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputGOBasedDEfromFeatures2.R
\name{OutputGOBasedDEfromFeatures2}
\alias{OutputGOBasedDEfromFeatures2}
\title{OutputGOBasedDEfromFeatures2}
\usage{
OutputGOBasedDEfromFeatures2(re.PJ.gene.based, gene.model, Output_file)
}
\arguments{
\item{Output_file}{}
}
\description{
Reformat genewised results from JunctionSeq, and use this one to get enriched GO terms
}
\examples{

RE.exon.sj.all.gene<-OutputGOBasedDEfromFeatures2(re.PJ.gene.based,gene.model,"GO_exon_sj_use_all_gene.xls")



}

