% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputGOBasedDEfromFeature3.R
\name{OutputGOBasedDEfromFeatures3}
\alias{OutputGOBasedDEfromFeatures3}
\title{OutputGOBasedDEfromFeatures3}
\usage{
OutputGOBasedDEfromFeatures3(re.PJ.gene.based, DE_type, gene.model,
  Output_file_dir)
}
\arguments{
\item{Output_file}{}
}
\description{
Reformat genewised results from JunctionSeq, and use this one to get enriched GO terms
}
\examples{

RE.exon.sj.all.gene<-OutputGOBasedDEfromFeatures3(re.PJ.gene.based,gene.model,"GO_exon_sj_use_all_gene.xls")



}

