% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwaysplice.R
\name{pathwaysplice}
\alias{pathwaysplice}
\title{pathwaysplice}
\usage{
pathwaysplice(pwf, genome, id, gene.model, gene2cat = NULL,
  test.cats = c("GO:CC", "GO:BP", "GO:MF"), method = "Wallenius",
  repcnt = 2000, use_genes_without_cat = FALSE)
}
\arguments{
\item{pwf}{probability weight function}

\item{genome}{genome you will use}

\item{id}{which gene id}

\item{gene.model}{which gene model}

\item{gene2cat}{a list with gene as name and category as value}

\item{test.cats}{the category you choose}

\item{method}{which method for calculting overpresentation p-value}

\item{repcnt}{the number of replications}

\item{use_genes_without_cat}{whether using genes without category or not}
}
\value{
The gene set enrichment results for gene sets
}
\description{
This function computes p-values for each gene set and output the list of 
differentially expressed genes within the gene set. 
The user can select serveral statistical methods for computing p-values
}
\examples{


}

