% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPWF_modified.R, R/plotPWF2.R
\name{plotPWF2}
\alias{plotPWF2}
\title{Plot the probability Weighting Function}
\usage{
plotPWF2(pwf, binsize = "auto", pwf_col = 3, pwf_lwd = 2,
  xlab = "Biased Data in <binsize> gene bins.", ylab = "Proportion DE", ...)

plotPWF2(pwf, binsize = "auto", pwf_col = 3, pwf_lwd = 2,
xlab = "Biased Data in <binsize> gene bins.", ylab = "Proportion DE", ...)
}
\arguments{
\item{pwf}{probability weigth function}

\item{binsize}{the number of gene in each bin(gene set)}

\item{pwf_col}{the color for the fitted line of probability weigth function}

\item{pwf_lwd}{the font for the fitted line of probability weigth function}

\item{xlab}{the label for x-axis}

\item{ylab}{the label for y-axis}

\item{...}{}

\item{pwf:}{The first step of the pathway analysis is to assess if there is any systematic bias of the significant genes. The plotPWF function takes input a vector of genes (with label 1 as significant and 0 otherwise), and the bias factor to be adjusted (e.g. number of splice junctions in the genes). The output of the function is a figure that illustrates the relationship between bias factor and proportion of significant genes. The size of the gene set size (i.e. number of genes in each gene bin) on the x-axis can be adjusted by user.}

\item{...}{}
}
\description{
Plot the probability Weighting Function
}
\examples{

png("~/GOSJ/Figure/pwfGeneGL.tiff")
plotPWF2(Gene.based.DE.feature.based.DE$pwfGeneGL)
dev.off()

png("~/GOSJ/Figure/pwfGeneFeature.tiff")
plotPWF2(Gene.based.DE.feature.based.DE$pwfGeneFeature)
dev.off()

png("~/GOSJ/Figure/pwfFeatureGL.tiff")
plotPWF2(Gene.based.DE.feature.based.DE$pwfFeatureGL)
dev.off()

png("~/GOSJ/Figure/pwFeatureFeature.tiff")
plotPWF2(Gene.based.DE.feature.based.DE$pwfFeatureFeature)
dev.off()


}

