% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{PostProcessGO}
\alias{PostProcessGO}
\title{PostProcessGO}
\usage{
PostProcessGO(n.go, adjusted, unadjuasted, venn.dir, boxplot.dir,
  In.ad.not.un.file, In.un.not.ad.file)
}
\arguments{
\item{n.go}{}

\item{adjusted}{}

\item{unadjuasted}{}

\item{venn.dir}{}

\item{boxplot.dir}{}

\item{In.ad.not.un.file}{}

\item{In.un.not.ad.file}{}
}
\value{
null
}
\description{
PostProcessGO
}

