% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessOutputFilesFrom_rMATS.R
\name{ProcessOutputFilesFrom_rMATS}
\alias{ProcessOutputFilesFrom_rMATS}
\title{ProcessOutputFilesFrom_rMATS}
\usage{
ProcessOutputFilesFrom_rMATS(dir.name, input.file.pattern, De_defined_by_what)
}
\arguments{
\item{input_file}{}
}
\description{
ProcessOutputFilesFrom_rMATS
}
\examples{

dir.name="/media/H_driver/PJ/rMAT/pHamard_MATS-selected-re-run-exon-centric/pHamard_MATS-selected-re-run-exon-centric/Neg_WTvKO_c0.01/"

input.file.pattern="*Neg_WTvKO_c0.01.txt"

sink(paste0(output.dir.name,"rMAT_output_from_PJ.txt"))
re.rMAT<-ProcessOutputFilesFrom_rMATS(dir.name,input.file.pattern)
sink()


}

