% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readJunctionSeqCounts.R
\name{readJunctionSeqCounts}
\alias{readJunctionSeqCounts}
\title{Title}
\usage{
readJunctionSeqCounts(countfiles = NULL, countdata = NULL, samplenames,
  design, flat.gff.file = NULL, test.formula1 = formula(~sample + countbin +
  condition:countbin), analysis.type = c("junctionsAndExons", "junctionsOnly",
  "exonsOnly"), nCores = 1, use.exons = NULL, use.junctions = NULL,
  use.known.junctions = TRUE, use.novel.junctions = TRUE,
  use.multigene.aggregates = FALSE, gene.names = NULL, verbose = TRUE,
  method.countVectors = c("geneLevelCounts", "sumOfAllBinsForGene",
  "sumOfAllBinsOfSameTypeForGene"), noDESeqMatrix = FALSE)
}
\arguments{
\item{noDESeqMatrix}{}
}
\description{
Title
}
\examples{
dir.name="/Volumes/Bioinformatics$/2015/Nimer_Cheng/"
file.sample="decoder.bySample.txt"
file.count="_junction_seq_new_gtf_7/QC.spliceJunctionAndExonCounts.forJunctionSeq.txt"
file.gff="Mus_musculus.GRCm38.83.JunctionSeq.flat.gff"
path.file.sample<-paste0(dir.name,file.sample)
decoder.bySample<-read.table(path.file.sample,header=T,stringsAsFactors = F)
print(decoder.bySample)
#Get count file
path.file.count<-paste0(dir.name,decoder.bySample$sample.ID,file.count)
countFiles<-paste0(path.file.count)
print(countFiles)
#Get annotation file
flat.file.gff<-paste0(dir.name,file.gff)
print(flat.file.gff)

method.countVectors = c("geneLevelCounts","sumOfAllBinsForGene","sumOfAllBinsOfSameTypeForGene")
method.countVectors <- match.arg(method.countVectors)

jscs = readJunctionSeqCounts(countfiles = as.character(sample.files),
samplenames = sample.names,
design = Re,
flat.gff.file = flat.gff.file,
 verbose = verbose,
 use.junctions = use.junctions,
 use.novel.junctions = use.novel.junctions,
 use.known.junctions = use.known.junctions,
 use.exons = use.exons,
 use.multigene.aggregates = use.multigene.aggregates,
 nCores = nCores,
 method.countVectors = method.countVectors,
 test.formula1 = test.formula1,
 gene.names = gene.names
 )



}

