% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReformatData.R
\name{ReformatData}
\alias{ReformatData}
\title{Title}
\usage{
ReformatData(re.PJ.gene.based)
}
\arguments{
\item{re.PJ.gene.based}{}
}
\description{
Title
}
\examples{
re.PJ.gene.based.testable.reformat<-ReformatData(re.PJ.gene.based)

}

