% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReformatDataUseTable.R
\name{ReformatDataUseTable}
\alias{ReformatDataUseTable}
\title{ReformatDataUseTable}
\usage{
ReformatDataUseTable(re.PJ.gene.based)
}
\arguments{
\item{re.PJ.gene.based}{}
}
\description{
ReformatDataUseTable
}
\examples{
re.PJ.gene.based.testable.reformat.2<-ReformatDataUseTable(Re.PJ.selected.feature.FC.p)

dim(re.PJ.gene.based.testable.reformat.2)

head(re.PJ.gene.based.testable.reformat.2)

}

