% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{reformatpath}
\alias{reformatpath}
\title{reformatpath}
\usage{
reformatpath(dir.name)
}
\arguments{
\item{dir.name}{Pathway name to be converted}
}
\description{
reformatpath
}
\examples{

dir.name <- '/media/H_driver/2016/Yang/MACS/MACS/'
reformatpath(dir.name)

}
