% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{Run_pathwaysplice}
\alias{Run_pathwaysplice}
\title{Run_pathwaysplice}
\usage{
Run_pathwaysplice(re.gene.based, ad = "GL", sub_feature = NULL, threshold,
  genomeID, geneID, gene_model, method)
}
\arguments{
\item{threshold}{}
}
\description{
Run_pathwaysplice
}
\examples{
data(mds)
data(hg19)

Example.Go.adjusted.by.exon<-Run_pathwaysplice(mds,ad="exon_SJ",sub_feature="E",0.05,genomeID="hg19",geneID="ensGene",gene_model=hg19.gene.model,method="Wallenius")

Example.Go.unadjusted<-Run_pathwaysplice(mds,ad="exon_SJ",sub_feature="E",0.05,genomeID="hg19",geneID="ensGene",gene_model=hg19.gene.model,method="Hypergeometric")

write.table(Example.Go.unadjusted[[1]]$DE_GO,file="DE.txt",col.names = F,row.names = F,quote=F)

}

