% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{runPathwaySplice}
\alias{runPathwaySplice}
\title{runPathwaySplice}
\usage{
runPathwaySplice(genewise.table, genome, id, gene2cat = NULL,
  test.cats = c("GO:CC", "GO:BP", "GO:MF"), go.size.cut = c(-Inf, Inf),
  method = "Wallenius", repcnt = 2000, use.genes.without.cat = FALSE,
  binsize = "auto")
}
\arguments{
\item{genewise.table}{data frame returned from function \code{makeGeneTable}}

\item{genome}{Genome to be used, options are 'hg19' or 'mm10'}

\item{id}{GeneID, options are 'entrezgene' or 'ensembl_gene_id'}

\item{gene2cat}{Get sets defined by users, can be obtained for example from \code{gmtGene2Cat} function}

\item{test.cats}{Default gene sets to be tested if \code{gene2cat} is not defined}

\item{go.size.cut}{Size limit of the gene sets to be tested}

\item{method}{the method used to calculate pathway enrichment p value. 
Options are 'Wallenius', 'Sampling', and 'Hypergeometric'}

\item{repcnt}{Number of random samples to be calculated when 'Sampling' is used
ignored unless \code{method='Sampling'}}

\item{use.genes.without.cat}{Whether genes not mapped to any category tested are included in analysis.
If set to FALSE, genes not mapped to any tested categories are ignored in analysis.}

\item{binsize}{The number of genes in each gene bin in the bias plot}
}
\value{
A dataframe with gene set enrichment analysis results
\item{geneID}{Gene ID}
\item{geneWisePvalue}{each gene is represented by the smallest p-value among its features}
\item{sig.gene}{a gene is significant (1) or not (0)} 
\item{mostSigDeFeature}{the most significant gene feature}
\item{numFeature}{number of gene features within the gene}
}
\description{
This function identifies pathways that are enriched with signficant genes, while accounting for 
different number of exons associated with each gene
}
\details{
This function implements the methodology described in Young et al. (2011) to adjust for 
         different number of counting bins associated with each gene. In the bias plot, the genes are grouped 
         by bias.factor into gene bins, the number of signficant genes are then plotted against the gene bins.
}
\examples{
res <- makeGeneTable(featureBasedData)
res <- runPathwaySplice(res,genome='hg19',id='ensGene',
                         test.cats=c('GO:BP'),
                         go.size.cut=c(5,30),
                         method='Wallenius',binsize=2)

 
}
\references{
Young MD, Wakefield MJ, Smyth GK, Oshlack A (2011) Gene ontology analysis for RNA-seq: 
accounting for selection bias. Genome Biology201011:R14
}
