% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select_DE_gene_based_on_Feature.R
\name{Select_DE_gene_basd_on_Feature}
\alias{Select_DE_gene_basd_on_Feature}
\title{Select_DE_gene_based_on_Feature}
\usage{
Select_DE_gene_basd_on_Feature(Re.PJ, re.PJ.gene.based, re.rMAT, splicing_type,
  cutoff_FC, cutoff_P_value, gene_model, venn_output)
}
\arguments{
\item{splicing_type:}{there are 5 splicing types from rMAT, need to choose one}
}
\description{
Select_DE_gene_based_on_Feature
}
\examples{

Re.PJ.selected.feature.FC.p<-Select_DE_gene_basd_on_Feature(Re.PJ,re.PJ.gene.based,re.rMAT,0.58,0.05,outputfile_DGE_FC_P_geneWise)
dim(Re.PJ.selected.feature.FC.p)
head(Re.PJ.selected.feature.FC.p)
length(which(Re.PJ.selected.feature.FC.p$DE_or_not==1))

Re.PJ.selected.feature.FC.p.check<-Select_DE_gene_basd_on_Feature(Re.PJ,re.PJ.gene.based,re.rMAT,"SE",2,0.05,outputfile_DGE_FC_P_geneWise)
}

