% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectFeature.R
\name{SelectFeature}
\alias{SelectFeature}
\title{SelectFeature}
\usage{
SelectFeature(Re.PJ)
}
\arguments{
\item{Re.PJ}{}
}
\description{
SelectFeature
}
\examples{
Re.PJ.selected<-SelectFeature(Re.PJ)

dim(Re.PJ.selected)
}

