% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationSJ2DE.R
\name{SimulationSJ2DE}
\alias{SimulationSJ2DE}
\title{Simulate gene sets based on the number of splicing junctions}
\usage{
SimulationSJ2DE(min_num_splicing, max_num_splicing, num_gene)
}
\arguments{
\item{min_num_splicing}{minium splicing junction number}

\item{max_num_splicing}{max splicing junction number}

\item{num_gene}{number of gene}
}
\value{
a list with two set of genes
}
\description{
We sample the mean of the number of splicing junction, and use this mean to simulate an array A including
splicing junction numbers with certain amounts. we convert this number into an array AAA between 0 and 1. Use the median of
AAA array as the probability for 1, we sample from [0,1] to get a list with certain length. This list is used as
a gene set. For example, if this gene set has 30 genes, each gene has its splicing junction number, we will
have 30 splicing junction number. After differential gene expression analysis, some genes are differentiallly expressed,
and are labeled as 1,otherwise labeled as 0.
Here we simulate two scenarios:
Scenario1: the probability of being 1 is dependent on the median of AAA
Scenario2: the probability of being 1 is not dependent on the median of AAA
}
\examples{
re.random.DE2SJ<-SimulationSJ2DE(20,200,30)
}

