% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{testPathwaySplice}
\alias{testPathwaySplice}
\title{testPathwaySplice}
\usage{
testPathwaySplice(featureBasedData, output.file.dir)
}
\arguments{
\item{featureBasedData}{A gene feature based table from differential analysis}

\item{output.file.dir}{Directory for output}
}
\value{
None
}
\description{
Peform one-step analysis for check bias, adjusted gene set 
enrichment anlysis and build network
}
\examples{

dir.name <- tempdir()
output.file.dir <- file.path(dir.name,'OutputTest')
testPathwaySplice(featureBasedData,output.file.dir = output.file.dir)


}
