% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UseLogistic2CKBias.R
\name{UseLogistic2CKBias}
\alias{UseLogistic2CKBias}
\title{UseLogistic2CKBias}
\usage{
UseLogistic2CKBias(jscs_genewise_object, genewise.pvalue = "geneWisePadj",
  sig.threshold = 0.05, type = c("exon", "splicing"))
}
\arguments{
\item{genewise.pvalue:}{whether you use smallest p value or geneWisePadj}

\item{sig.threshold:}{threshold to define differential gene list}

\item{type:}{whether you are interested in exon or splicing junction}

\item{jscs_genewise_object:}{an object returned from ReformatData and MakeGeneWiseTable_JunctionSeq}
}
\description{
UseLogistic2CKBias
}
\examples{
re.PJ.gene.based.testable.reformat<-ReformatData(re.PJ.gene.based)
UseLogistic2CKBias(re.PJ.gene.based.testable.reformat)

data(mds)
re<-UseLogistic2CKBias(mds)

}

