% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UseOtherDataSet.R
\name{UseOtherDataSet}
\alias{UseOtherDataSet}
\title{Title}
\usage{
UseOtherDataSet(input_file)
}
\arguments{
\item{input_file}{}
}
\description{
Title
}
\examples{




Re<-UseOtherDataSet("/media/H_driver/DataSet_SJ/GSE66793_cmp1-3-geo-juncs.tsv")


}

