% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callQoRT.R
\name{callQoRT}
\alias{callQoRT}
\title{Title}
\usage{
callQoRT(cmd1, inputfile, outfile)
}
\arguments{
\item{outfile}{}
}
\description{
Title
}
\examples{
callQoRT("java -Xmx4000M -jar /home/aiminyan/QoRTs/QoRTsFullExampleData/QoRTsRelease/QoRTs.jar QC --singleEnded --noGzipOutput --maxReadLength 1000 --keepMultiMapped",
         "/media/H_driver/2015/Nimer_Cheng/164.alignments.bam /media/DATA/mus_musculus/Mus_musculus.GRCm38.83.processed.sorted.gtf",
         "/media/H_driver/2015/Nimer_Cheng/164_junction_seq_new_gtf_7")
}

