% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckBias.R
\name{Cbs}
\alias{Cbs}
\title{Cbs}
\usage{
Cbs(re.gene.based, ad = "GL", sub_feature = NULL, threshold, genomeID,
  geneID, gene_model, method)
}
\arguments{
\item{re.gene.based}{}

\item{ad}{}

\item{sub_feature}{}

\item{threshold}{}

\item{genomeID}{}

\item{geneID}{}

\item{gene_model}{}

\item{method}{}
}
\description{
Cbs is to ckeck possible bias factor using the method in goseq
}
\examples{

Cbs(mds,ad="E",sub_feature="E",threshold=0.05)


}

