% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckBias.R
\name{Cbs}
\alias{Cbs}
\title{Cbs}
\usage{
Cbs(re.gene.based, adjust = "GL", sub_feature = NULL, threshold, genomeID,
  geneID)
}
\arguments{
\item{re.gene.based}{Gene based table converted from DU results}

\item{adjust}{The possible bias factor to be used(gene length or number of exons)}

\item{sub_feature}{use exon("E") or splicing junction("J") to get gene table}

\item{threshold}{cutoff for perGeneQvalue calculated using the method in DEXSeq}

\item{genomeID}{which genome(mm10 or hg19)}

\item{geneID}{which type of gene ID("ensGene" or "geneSymbol")}
}
\value{
A data frame that includes gene ID, status of differential gene
and probability weight function
}
\description{
Cbs is to check possible bias factor using the method in goseq
}
\examples{

data(mds)
Cbs(mds,adjust='E',sub_feature='E',threshold=0.05)


}
