% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{compareResults}
\alias{compareResults}
\title{compareResults}
\usage{
compareResults(n.go, adjusted, unadjusted, gene.based.table,
  output.dir = tempdir(), type.boxplot = c("All", "Only3"))
}
\arguments{
\item{n.go}{Number of gene sets}

\item{adjusted}{An object returned by \code{runPathwaySplice}, should correspond to
gene set anlaysis results adjusting for biases in splicing analysis}

\item{unadjusted}{An object returned by \code{runPathwaySplice}, should correspond to 
gene set analysis results NOT adjusting for biases}

\item{gene.based.table}{An object returned by \code{makeGeneTable}, should correspond to 
a table with one p-value for each gene}

\item{output.dir}{Directory for output files}

\item{type.boxplot}{Options are "All" and "Only3", corresponding to drawing 5 boxplots or 3 boxplots. 

       5 boxplots:  \code{all genesets}, \code{sig.adjusted} (sig gene sets in adjusted analysis),
        \code{sig.unadjusted} (sig gene sets in unadjusted analysis),
        \code{sig.adjusted.only} (sig gene sets in adjusted analysis only), 
        \code{sig.unadjusted.only} (sig gene sets in unadjusted analysis only)
                     
       3 boxplots: \code{all genesets}, \code{adjusted.sig}, \code{unadjusted.sig}}
}
\value{
The output include 3 files in \code{output.dir}: 
(1) a venn diagram comparing significant gene sets before and after adjusting for bias factors
(2) a .csv file with gene set names belonging to different sections of the venn diagram    
(3) a box plot showing the distributions of number of features within all genes 
in significant gene sets, with and without adjusting for bias factors
}
\description{
This function compares the distributions of bias factors (e.g. number of exon bins) 
in genes within significant gene sets, with and without adjusting for bias factors 
in splicing pathway analysis.
}
\examples{

dir.name <- system.file('extdata', package='PathwaySplice')
hallmark.pathway.file <- file.path(dir.name,"h.all.v6.0.symbols.gmt.txt")

hallmark <- gmtGene2Cat(hallmark.pathway.file,genomeID='hg19')
                   
gene.based.table <- makeGeneTable(featureBasedData)

res.adj <- runPathwaySplice(gene.based.table,genome='hg19',
                         id='ensGene',gene2cat=hallmark,  
                         go.size.limit = c(5, 200),
                         method='Wallenius', output.file=tempfile())

res.unadj <- runPathwaySplice(gene.based.table,genome='hg19',
                         id='ensGene',gene2cat=hallmark,go.size.limit = c(5, 200),
                         method='Hypergeometric',output.file=tempfile())

compareResults(20, res.adj, res.unadj, gene.based.table, type.boxplot='Only3')


}
