% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{compareResults}
\alias{compareResults}
\title{compareResults}
\usage{
compareResults(n.go, adjusted, unadjusted, output.dir, type.boxplot = c("All",
  "Only3"), In.ad.not.un.file, In.un.not.ad.file)
}
\arguments{
\item{n.go}{Number of gene sets}

\item{adjusted}{Adjusted result}

\item{unadjusted}{Unadjusted result}

\item{output.dir}{Path for output( including venn,boxplot and two tables)}

\item{type.boxplot}{Get boxplot for 5 categories or 3 categories

       6 categories: 'All.adjusted','All.unadjusted',
                     'Top25.adjusted','Top25.unadjusted',
                     'In_ad_not_un','In_un_not_ad'
                     
       3 categories: 'All','Top25.adjusted','Top25.unadjusted}

\item{In.ad.not.un.file}{File name for outputing adjused but not
in unadjusted when using the selected gene sets}

\item{In.un.not.ad.file}{File name for outputing unadjused but not
in adjusted when using the selected gene sets}
}
\value{
Nothing to be returned
}
\description{
compareResults
}
\examples{

dir.name <- system.file('extdata', package='PathwaySplice')
canonical.pathway.file <- file.path(dir.name,"h.all.v6.0.symbols.gmt.txt")

cpp <- gmtGene2Cat(canonical.pathway.file,genomeID='hg19')
                   
gene.based.table <- makeGeneTable(featureBasedData)

res1 <- runPathwaySplice(gene.based.table,genome='hg19',
                         id='ensGene',gene2cat=cpp,go.size.limit = c(2, 200),
                         method='Wallenius')

res2 <- runPathwaySplice(gene.based.table,genome='hg19',
                         id='ensGene',gene2cat=cpp,go.size.limit = c(2, 200),
                         method='Hypergeometric')

output.dir <- file.path(tempdir(),'OutputPostAnalysis')

output.file.name.1 <- 'In_ad_not_un.xls'
output.file.name.2 <- 'In_un_not_ad.xls'

compareResults(4,res1,res2,output.dir,
                      type.boxplot='Only3',
                      output.file.name.1,output.file.name.2)
}
