% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{compareResults}
\alias{compareResults}
\title{Compares distribution of xx with and without adjusting for bias factors in splicing pathway analysis.}
\usage{
compareResults(n.go, adjusted, unadjusted, output.dir = tempdir(),
  type.boxplot = c("All", "Only3"), In.ad.not.un.file = tempfile(),
  In.un.not.ad.file = tempfile())
}
\arguments{
\item{n.go}{Number of gene sets}

\item{adjusted}{An object returned by \code{runPathwaySplice}, should correspond to
gene set anlaysis results adjusting for biases in splicing analysis}

\item{unadjusted}{An object returned by \code{runPathwaySplice}, should correspond to 
gene set analysis results NOT adjusting for biases.}

\item{output.dir}{Directory for output files}

\item{type.boxplot}{Get boxplot for 5 categories or 3 categories

       6 categories: 'All.adjusted','All.unadjusted',
                     'Top25.adjusted','Top25.unadjusted',
                     'In_ad_not_un','In_un_not_ad'
                     
       3 categories: 'All','Top25.adjusted','Top25.unadjusted}

\item{In.ad.not.un.file}{File name for outputing adjused but not
in unadjusted when using the selected gene sets}

\item{In.un.not.ad.file}{File name for outputing unadjused but not
in adjusted when using the selected gene sets}
}
\value{
The output include 3 files in \code{output.dir}: 
(1) a venn diagram comparing significant gene sets before and after adjusting for bias factors
(2) a box plot xxx 
(3) xxx
}
\description{
Compares distribution of xx with and without adjusting for bias factors in splicing pathway analysis.
}
\examples{

dir.name <- system.file('extdata', package='PathwaySplice')
hallmark.pathway.file <- file.path(dir.name,"h.all.v6.0.symbols.gmt.txt")

hallmark <- gmtGene2Cat(hallmark.pathway.file,genomeID='hg19')
                   
gene.based.table <- makeGeneTable(featureBasedData)

res1 <- runPathwaySplice(gene.based.table,genome='hg19',
                         id='ensGene',gene2cat=hallmark,  
                         go.size.limit = c(2, 200),
                         method='Wallenius', output.file=tempfile())

res2 <- runPathwaySplice(gene.based.table,genome='hg19',
                         id='ensGene',gene2cat=hallmark,go.size.limit = c(2, 200),
                         method='Hypergeometric',output.file=tempfile())

compareResults(50,res1,res2,tempdir(),type.boxplot='Only3')

}
