% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enMap2.R
\name{enMap2}
\alias{enMap2}
\title{enMap2}
\usage{
enMap2(GoSeqRes, gene.set.type = "GO", n = 50, fixed = TRUE,
  vertex.label.font = 1, SimilarityThreshold, ...)
}
\arguments{
\item{GoSeqRes}{object returned from Run_pathwaysplice}

\item{gene.set.type}{whether you are interested in GO, KEGG, or other pathways}

\item{n}{maximum number of category to shown}

\item{fixed}{if set to FALSE, will invoke tkplot}

\item{vertex.label.font}{font size of vertex label}

\item{SimilarityThreshold}{threshold for defining similarity between GOs}

\item{...}{additional parameter}
}
\value{
A figure for visualizing network
}
\description{
enMap2 is used to draw network based on similarities between GOs
}
\author{
Aimin created this funciton based on enrichMap function in G Yu's DOSE R package
}

