% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enMap.R
\name{enrichmentMap}
\alias{enrichmentMap}
\title{enrichmentMap}
\usage{
enrichmentMap(GoSeqRes, n = 50, fixed = TRUE, vertex.label.font = 1,
  SimilarityThreshold, ...)
}
\arguments{
\item{GoSeqRes}{Object returned from Run_pathwaysplice}

\item{n}{Maximum number of category to be shown}

\item{fixed}{If set to FALSE, will invoke tkplot}

\item{vertex.label.font}{Font size of vertex label}

\item{SimilarityThreshold}{Threshold for defining Jaccard Coefficient(JC)
JC ranges from 0 to 1:
JC=0, indicates there are no overlap on genes between
       two gene sets
JC=1, indicates two gene sets are identical  
SimilarityThreshold=0, indicates the enrichment map includes
all gene sets with their mutual JC greater than 0}

\item{...}{Additional parameter}
}
\value{
A figure for visualizing enrichment network
}
\description{
enrichmentMap is used to draw network based on similarities between GOs
}
\examples{

res <- Run_pathwaysplice(mds.11.sample,adjust='exon_SJ',sub_feature='E',
0.05,genomeID='hg19',geneID='ensGene', method='Wallenius')

enmap<-enrichmentMap(res,n=5,SimilarityThreshold=0)
}
\author{
Aimin created this funciton based on enrichMap function in G Yu's DOSE R package
}
