% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enMap2.R
\name{enrichmentMap}
\alias{enrichmentMap}
\title{enrichmentMap}
\usage{
enrichmentMap(GoSeqRes, gene.set.type = "GO", n = 50, fixed = TRUE,
  vertex.label.font = 1, SimilarityThreshold, ...)
}
\arguments{
\item{GoSeqRes}{object returned from Run_pathwaysplice}

\item{gene.set.type}{whether you are interested in GO, KEGG, or other pathways}

\item{n}{maximum number of category to shown}

\item{fixed}{if set to FALSE, will invoke tkplot}

\item{vertex.label.font}{font size of vertex label}

\item{SimilarityThreshold}{threshold for defining similarity between GOs}

\item{...}{additional parameter}
}
\value{
A figure for visualizing network
}
\description{
enrichmentMap is used to draw network based on similarities between GOs
}
\examples{

data(mds)

Example.Go.adjusted.by.exon<-Run_pathwaysplice(mds,ad="exon_SJ",
sub_feature="E",0.05,genomeID="hg19",geneID="ensGene",
gene_model=hg19,method="Sampling")

re.w.adjusted<-enrichmentMap(Example.Go.adjusted.by.exon,n=5,SimilarityThreshold=0)
Example.Go.unadjusted<-Run_pathwaysplice(mds,ad="exon_SJ",
sub_feature="E",0.05,genomeID="hg19",geneID="ensGene",
gene_model=hg19,method="Hypergeometric")

re.w.unadjusted<-enrichmentMap(Example.Go.unadjusted,n=5,SimilarityThreshold=0)



}
\author{
Aimin created this funciton based on enrichMap function in G Yu's DOSE R package
}

