% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{enrichmentMap}
\alias{enrichmentMap}
\title{enrichmentMap}
\usage{
enrichmentMap(goseqres, n = 50, fixed = TRUE, vertex.label.font = 1,
  similarity.threshold, output.file.dir = tempdir(),
  label.vertex.by.index = FALSE, ...)
}
\arguments{
\item{goseqres}{Object returned from runPathwaySplice}

\item{n}{The top \emph{n} categories are shown in enrichment map}

\item{fixed}{If set to FALSE, will invoke tkplot to let user to plot enrichment map 
using an interactive graph drawing facility in R. Note: user needs to 
have XQuartz and tcltk be install}

\item{vertex.label.font}{Font size of vertex label}

\item{similarity.threshold}{Gene sets with Jaccard Coefficient > similarity.threshold 
will be connected on the enrichment map}

\item{output.file.dir}{Output dir for the gene set information file on network. 
User can find a negtwork file in GML format in this directory that can be   #'                       used as an input for Cytoscape}

\item{label.vertex.by.index}{Which way to be used for labeling vertex on network

FALSE indicates to label vertex by the name of gene sets

TRUE indicates to label vertex by the index of gene sets}

\item{...}{Additional parameter}
}
\value{
A list for giving edge and vertex information of enrichment map
}
\description{
This function draws an enrichment map using the similarities 
between gene sets measured by Jaccard Coefficient(JC)
}
\details{
In the enrichment map, vertex color corresponds to over represented pvalue;
vertex size corresponds to the number of 'significant' gene in gene set;
edge width corresponds to Jaccard similarity coefficient.

The Jaccard similarity coefficient ranges from 0 to 1. JC=0 indicates 
there are no overlapping genes between two gene sets, 
JC=1 indicates two gene sets are identical.
}
\examples{

gene.based.table <- makeGeneTable(featureBasedData)

res <- runPathwaySplice(gene.based.table,genome='hg19',
                         id='ensGene',test.cats=c('GO:BP'),
                         go.size.limit=c(5,30),
                         method='Wallenius')

}
\author{
Aimin created this funciton based on enrichMap function in G Yu's DOSE R package
}
