% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{enrichmentMap}
\alias{enrichmentMap}
\title{enrichmentMap}
\usage{
enrichmentMap(goseqres, n = 50, fixed = TRUE, vertex.label.font = 1,
  similarity.threshold, output.file.dir, label.vertex.by.index = FALSE, ...)
}
\arguments{
\item{goseqres}{Object returned from runPathwaySplice}

\item{n}{Maximum number of category to be shown}

\item{fixed}{If set to FALSE, will invoke tkplot}

\item{vertex.label.font}{Font size of vertex label}

\item{similarity.threshold}{Threshold for defining Jaccard Coefficient(JC)

JC ranges from 0 to 1:

JC=0, indicates there are no overlap on genes between
       two gene sets

JC=1, indicates two gene sets are identical  

similarity.threshold=0, indicates the enrichment map includes
all gene sets with their mutual JC greater than 0}

\item{output.file.dir}{Output dir for the gene set information file on network}

\item{label.vertex.by.index}{Which way to be used for labeling vertex on network

FALSE indicates to label vertex by the name of gene sets

TRUE indicates to label vertex by the index of gene sets}

\item{...}{Additional parameter}
}
\value{
A list for giving edge and vertex information of enrichment map
}
\description{
enrichmentMap is used to draw Enrichment Map based on similarities defined using Jaccard Coefficient between GOs or gene sets
}
\examples{

gene.based.table <- makeGeneTable(featureBasedData)

res <- runPathwaySplice(gene.based.table,genome='hg19',
                         id='ensGene',test.cats=c('GO:BP'),
                         go.size.cut=c(5,30),
                         method='Wallenius')

output.file.dir <- file.path(tempdir(),'OutputEnmapEx')

enmap <- enrichmentMap(res,n=10,similarity.threshold=0,
                       output.file.dir = output.file.dir,
                       label.vertex.by.index = TRUE)
                       
}
\author{
Aimin created this funciton based on enrichMap function in G Yu's DOSE R package
}
