% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc4Data.R
\docType{data}
\name{Feature10000BasedData}
\alias{Feature10000BasedData}
\title{Feature10000BasedData}
\format{A data frame with 10000 rows (exons) and 3 columns (variables)}
\usage{
data(Feature10000BasedData)
}
\value{
\item{geneID}{Gene ID}
\item{countbinID}{ID for features of each gene (e.g. exons or splicing junctions)}
\item{pvalue}{p-value for differential usage analysis of the features defined by 
      \code{countbinID}}
}
\description{
A dataset generated by sampling 10000 exons from differential usage analysis results of 
RNA-seq data from Dolatshad et al. (2015)
}
\references{
H Dolatshad, A Pellagatti, M Fernandez-Mercado1, B H Yip, L Malcovati, M Attwood, B Przychodzen
N Sahgal, A A Kanapin, H Lockstone, L Scifo, P Vandenberghe, E Papaemmanuil, C W J Smith, P J Campbell, 
S Ogawa1, J P Maciejewski, M Cazzola, K I Savage1 and J Boultwood1 (2015) \emph{Disruption of SF3B1 results in deregulated 
expression and splicing of key genes and pathways in myelodysplastic syndrome hematopoietic stem and progenitor 
cells.} Leukemia (2015) 29, 1092-1103
}
\keyword{datasets}
