% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc4Data.R
\docType{data}
\name{gene.model}
\alias{gene.model}
\title{mm10 Gene annotation data set}
\format{A data frame with 23440 rows and 6 variables}
\usage{
data(gene.model)
}
\value{
a data set 23440 rows and 6 variables
}
\description{
a data set for mm10 gene annotation
}
\keyword{datasets}

