% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Over_under_represented_pvalue.R
\name{Get_Over_under_represented_pvalue}
\alias{Get_Over_under_represented_pvalue}
\title{calculate the p-value for GO terms}
\usage{
Get_Over_under_represented_pvalue(num_de_incat, num_incat, num_genes, num_de,
  weight)
}
\arguments{
\item{weight}{}
}
\description{
calculate the p-value for GO terms
}
\examples{

28 	 186 	 1852 	 125 	 1.03043

use the gene pooled from all GO terms
Get_Over_under_represented_pvalue(28,186,1852,125,1.03043)

use the genes that have matched GO terms

Get_Over_under_represented_pvalue(28,186,1840,125,1.026767)

}

