% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdaptedCodes_juncbias.R
\name{getDataset}
\alias{getDataset}
\title{Simulation studies}
\usage{
getDataset(counts, drop.extreme.dispersion = 0.1, drop.low.lambda = TRUE)
}
\arguments{
\item{drop.low.lambda}{}
}
\description{
Simulation studies
}

