% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgo3.R
\name{getgo3}
\alias{getgo3}
\title{getgo3}
\usage{
getgo3(genes, genome, id, fetch.cats = c("GO:CC", "GO:BP", "GO:MF"))
}
\arguments{
\item{genes}{gene list}

\item{genome}{genome that is used}

\item{id}{Gene id}

\item{fetch.cats}{Gene sets that you are interested}
}
\value{
a gene set(GO or pathway)
}
\description{
Based on a gene list, and return a gene set that these genes belong to
}
\examples{

re<-getgo3(rownames(Re.Go.adjusted.by.number.junction.2[[2]]),"mm10","ensGene",fetch.cats=c("GO:BP"))

}

