% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgo3.R
\name{getgo3}
\alias{getgo3}
\title{Title}
\usage{
getgo3(genes, genome, id, fetch.cats = c("GO:CC", "GO:BP", "GO:MF"))
}
\arguments{
\item{fetch.cats}{}
}
\description{
Title
}
\examples{

getgo3(rownames(Re.Go.adjusted.by.number.junction.2[[2]]),"mm10","ensGene",fetch.cats=c("GO:BP"))

}

