% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPwfUseReformatedData.R
\name{GetPwfUseReformatedData}
\alias{GetPwfUseReformatedData}
\title{GetPwfUseReformatedData}
\usage{
GetPwfUseReformatedData(re.gene.based, ad = "GL", sub_feature = NULL,
  threshold)
}
\arguments{
\item{threshold}{}
}
\description{
Use reformated data to calculate probability weight function
}
\examples{
Re.pwf.exon.sj<-GetPwfUseReformatedData(ReformatData(re.PJ.gene.based),ad="exon_SJ",sub_feature=NULL,0.05)

Re.pwf.exon.sj<-GetPwfUseReformatedData(mds,ad="exon_SJ",sub_feature="E",0.05)

}

