% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_get_results_from_JunctionSeq.R
\name{GetResultsFromJunctionSeq}
\alias{GetResultsFromJunctionSeq}
\title{Get analysis results using JunctionSeq}
\usage{
GetResultsFromJunctionSeq(dir.name, file.sample, file.count, file.gff)
}
\arguments{
\item{file.gff}{}
}
\description{
Get analysis results using JunctionSeq
}
\examples{

load("/Volumes/Bioinformatics\\$/2015/Nimer_Cheng/1_29_2016.RData")

dir.name="/Volumes/Bioinformatics$/2015/Nimer_Cheng/"
file.sample="decoder.bySample.txt"
file.count="_junction_seq_new_gtf_7/QC.spliceJunctionAndExonCounts.forJunctionSeq.txt"
file.gff="Mus_musculus.GRCm38.83.JunctionSeq.flat.gff"
Re<-GetResultsFromJunctionSeq(dir.name,file.sample,file.count,file.gff)
head(fData(Re))
save(Re,file="Re_Run_test_GOSJ.RData")

dir.name.PJ="/media/H_driver/PJ/"
file.sample.PJ="decoder.bySample.rtf"
file.count.PJ="/QC.spliceJunctionAndExonCounts.forJunctionSeq.txt"

file.gff=paste0("/media/H_driver/2015/Nimer_Cheng/",file.gff)
Re.PJ<-GetResultsFromJunctionSeq(dir.name.PJ,file.sample.PJ,file.count.PJ,file.gff)

save(Re.PJ,file="/media/H_driver/PJ/PJ_jscs.RData")


buildAllPlots(jscs=jscs,outfile.prefix="./plots_based_on_DE_splice_site_gene1/",
gene.list=gene.based.de.splice.site[1],use.plotting.device="png",plot.gene.level.expression=TRUE,sequencing.type="single-end");





}

