% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_get_results_from_JunctionSeq.R
\name{GetResultsFromJunctionSeq}
\alias{GetResultsFromJunctionSeq}
\title{Get analysis results using JunctionSeq}
\usage{
GetResultsFromJunctionSeq(dir.name, file.sample, file.count, file.gff)
}
\arguments{
\item{file.gff}{}
}
\description{
Get analysis results using JunctionSeq
}
\examples{

load("/Volumes/Bioinformatics\\$/2015/Nimer_Cheng/1_29_2016.RData")
dir.name="/Volumes/Bioinformatics$/2015/Nimer_Cheng/"
file.sample="decoder.bySample.txt"
file.count="_junction_seq_new_gtf_7/QC.spliceJunctionAndExonCounts.forJunctionSeq.txt"
file.gff="Mus_musculus.GRCm38.83.JunctionSeq.flat.gff"
Re<-GetResultsFromJunctionSeq(dir.name,file.sample,file.count,file.gff)
head(fData(Re))

save(Re,file="Re_Run_test_GOSJ.RData")

}

