% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{getResultsFromJunctionSeq}
\alias{getResultsFromJunctionSeq}
\title{getResultsFromJunctionSeq}
\usage{
getResultsFromJunctionSeq(dir.name, sample.file, count.file, gff.file,
  method.dispFinal = c("shrink", "max", "fitted", "noShare"))
}
\arguments{
\item{dir.name}{Path name for sample information file}

\item{sample.file}{Sample information file}

\item{count.file}{Count file}

\item{gff.file}{Annotation file}

\item{method.dispFinal}{Determine the method 
used to get a 'final' dispersion estimate.}
}
\value{
The analysis result from JunctionSeq R package
}
\description{
This function is used to get analysis results from using JunctionSeq
}
\examples{


dir.name <- system.file('extdata', package='PathwaySplice')
sample.file <- 'Sample_info.txt'
count.file <- 'Counts.10.genes.txt'
gff.file <- 'flat.chr22.10.genes.gff'
res <- getResultsFromJunctionSeq(dir.name, sample.file, 
                                 count.file,gff.file, 
                                 method.dispFinal = 'max')

}
