% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_get_results_from_JunctionSeq.R
\name{GetResultsFromJunctionSeq}
\alias{GetResultsFromJunctionSeq}
\title{Get analysis results using JunctionSeq}
\usage{
GetResultsFromJunctionSeq(dir.name, file.sample, file.count, file.gff)
}
\arguments{
\item{file.gff}{}
}
\description{
Get analysis results using JunctionSeq
}
\examples{

# For example data set

dir.name="/media/H_driver/Aimin_project/"

file.sample="decoder.bySample.Mut_WT_2.rtf"
file.gff="Homo_sapiens.GRCh38.84.processed.sorted.4.JunctionSeq.flat.gff"
file.count="/QC.spliceJunctionAndExonCounts.forJunctionSeq.txt"

file.gff=paste0(dir.name,"GTF_Files/",file.gff)

Re.example<-GetResultsFromJunctionSeq(dir.name,file.sample,file.count,file.gff)
save(Re.example,file=paste0(getwd(),"/data/example_Mut_WT.RData"))

file.sample.2="decoder.bySample.Mut_HC_2.rtf"
Re.example.Mut.HC<-GetResultsFromJunctionSeq(dir.name,file.sample.2,file.count,file.gff)

buildAllPlots(jscs=jscs,outfile.prefix="./plots_based_on_DE_splice_site_gene1/",
gene.list=gene.based.de.splice.site[1],use.plotting.device="png",plot.gene.level.expression=TRUE,sequencing.type="single-end");



}

