% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_get_results_from_JunctionSeq.R
\name{GetResultsFromJunctionSeq}
\alias{GetResultsFromJunctionSeq}
\title{GetResultsFromJunctionSeq
Get analysis results using JunctionSeq}
\usage{
GetResultsFromJunctionSeq(dir.name, file.sample, file.count, file.gff)
}
\arguments{
\item{dir.name}{path name for sample information file}

\item{file.sample}{sample information file}

\item{file.count}{count file}

\item{file.gff}{annotation file}
}
\value{
return an analysis results from JunctionSeq
}
\description{
GetResultsFromJunctionSeq
Get analysis results using JunctionSeq
}
\examples{
file.sample="Sample_info.txt"
dir.name=dirname(system.file("extdata","Sample_info.txt", package = "PathwaySplice"))
dir.name=paste0(dir.name,"/")
file.gff="flat.chr22.gff"
file.count="/QC.spliceJunctionAndExonCounts.forJunctionSeq.txt"
\donttest{Re.example<-GetResultsFromJunctionSeq(dir.name,file.sample,file.count,file.gff)}

}
