% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gmt2GeneCat.R
\name{Gmt2GeneCat}
\alias{Gmt2GeneCat}
\title{Read a gmt file, return a list with the name of element being a gene id based on gene_anno_file, and each element
being the pathways that this gene corresponds to}
\usage{
Gmt2GeneCat(gmt_input_file, gene_anno_file)
}
\arguments{
\item{gene_anno_file}{}
}
\description{
Read a gmt file, return a list with the name of element being a gene id based on gene_anno_file, and each element
being the pathways that this gene corresponds to
}

