% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gmt2GeneCat.R
\name{Gmt2GeneCat}
\alias{Gmt2GeneCat}
\title{Gmt2GeneCat}
\usage{
Gmt2GeneCat(gmt_input_file, file.type, gene_anno_file = NULL, gene_anno)
}
\arguments{
\item{gmt_input_file}{Input file}

\item{file.type}{Local or url}

\item{gene_anno_file}{Gene annotation file supplied as a file}

\item{gene_anno}{Gene annotation data set(mm10 or hg19 in this package) to be used if no annotation file}
}
\value{
a list with its names being geneID, its elements being the pathways
}
\description{
Read a gmt file, and return a list with its name
being a gene id based on gene_anno_file, and each element of this list
being the pathways that this gene corresponds to
}
\examples{

cp.gmt.file=system.file("extdata","c2.cp.v5.2.symbols.gmt.txt", package = "PathwaySplice")
data(hg38)

\donttest{gene.2.cat.hallmark.hg<-Gmt2GeneCat(cp.gmt.file,'local',gene_anno=hg38)}

}
