% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gmt2GeneCat.R
\name{Gmt2GeneCat}
\alias{Gmt2GeneCat}
\title{Gmt2GeneCat}
\usage{
Gmt2GeneCat(gmt_input_file, file.type, gene_anno_file = NULL, gene_anno)
}
\arguments{
\item{gmt_input_file}{input file}

\item{file.type}{local or url}

\item{gene_anno_file}{annotation file}

\item{gene_anno}{gene annotation to be used if no annotation file}
}
\value{
a list with its names being geneID, its element being the pathways
}
\description{
Read a gmt file, and return a list with its name
being a gene id based on gene_anno_file, and each element of this list
being the pathways that this gene corresponds to
}
\examples{

cp.gmt.file=system.file("extdata","c2.cp.v5.2.symbols.gmt.txt", package = "PathwaySplice")
data(hg38)

\donttest{gene.2.cat.hallmark.hg<-Gmt2GeneCat(cp.gmt.file,'local',gene_anno=hg38)}

}

