% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{gmtGene2Cat}
\alias{gmtGene2Cat}
\title{gmtGene2Cat}
\usage{
gmtGene2Cat(dir.name, pathway.file, location.type, gene.anno.file = NULL,
  genomeID = c("mm10", "hg19", "hg38"))
}
\arguments{
\item{dir.name}{Directory where the gene sets file (in GMT format) is located}

\item{pathway.file}{Input file for the gene sets in GMT format}

\item{location.type}{Location of the gene set file. Valid options are "local" or "url"}

\item{gene.anno.file}{Gene annotation file supplied as a file}

\item{genomeID}{Genome ('mm10','hg19' or 'hg38') to be used}
}
\value{
A list where each entry is named by a gene and contains a vector of all
        the pathways associated with the gene
}
\description{
Obtains all pathways associated with a set of genes
}
\details{
This function reads a gene set file in GMT format, and returns a list with its name
being a gene id, and each element of the list being the pathways associated with the gene
}
\examples{

dir.name <- system.file('extdata', package='PathwaySplice')
canonical.pathway.file <- '10.cp.gmt.txt'
res <- gmtGene2Cat(dir.name,canonical.pathway.file,'local',genomeID='hg19')

}
