% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{gmtGene2Cat}
\alias{gmtGene2Cat}
\title{gmtGene2Cat}
\usage{
gmtGene2Cat(dir.name, pathway.file, file.type, gene.anno.file = NULL,
  genomeID = c("mm10", "hg19", "hg38"))
}
\arguments{
\item{dir.name}{Directory for the gene sets in GMT format that is located in}

\item{pathway.file}{Input file for the gene sets in GMT format}

\item{file.type}{Indicates the gene set file in GMT format is in local or url}

\item{gene.anno.file}{Gene annotation file supplied as a file}

\item{genomeID}{Genome ('mm10','hg19' or 'hg38') to be used}
}
\value{
A list with its names being geneID, its elements being the pathways
}
\description{
Read a gene set file in GMT format, and return a list with its name
being a gene id, and each element of this list
being the pathways that this gene corresponds to
}
\examples{

dir.name <- system.file('extdata', package='PathwaySplice')
canonical.pathway.file <- '10.cp.gmt.txt'
res <- gmtGene2Cat(dir.name,canonical.pathway.file,'local',genomeID='hg19')

}
