% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{gmtGene2Cat}
\alias{gmtGene2Cat}
\title{gmtGene2Cat}
\usage{
gmtGene2Cat(pathway.file, gene.anno.file = NULL, genomeID = c("mm10",
  "hg19", "hg38"))
}
\arguments{
\item{pathway.file}{Input file for the gene sets in GMT format, must be in \emph{gene symbols}}

\item{gene.anno.file}{Gene annotation file that facilitate gene id conversions when 
gene ids in RNA-Seq data and \code{pathway.file} differ. If not specified, \code{gmtGene2cat} 
relies on gene annotations provided by R package \code{AnnotationHub}.}

\item{genomeID}{Genome to be used. Options are 'mm10','hg19' or 'hg38'.}
}
\value{
A list where each entry is named by a gene and contains a vector of all
        the pathways associated with the gene
}
\description{
Obtains all pathways associated with a set of genes
}
\details{
This function reads a gene set file in \url{https://software.broadinstitute.org/cancer/
software/gsea/wiki/index.php/Data_formats#GMT:_Gene_Matrix_Transposed_file_format_.28.2A.gmt.29}{GMT format},
and returns a list with its name being a gene id, and each element of 
the list being the pathways associated with the gene. When gene ids in RNA-Seq data differ from those in pathway database,
\code{gene.anno.file} facilitate gene id conversions. Users can prepare this file based on the format of the example gene annotation file at 
             \url{https://raw.githubusercontent.com/aiminy/GOSJ/master/data/gene_annotation.txt}{this link.}
}
\examples{
#using local file for pathways database
dir.name <- system.file('extdata', package='PathwaySplice')
hallmark.local.pathways <- file.path(dir.name,'h.all.v6.0.symbols.gmt.txt')
hlp <- gmtGene2Cat(hallmark.local.pathways, genomeID='hg19')

\dontrun{
#using url for pathways database linked to a website
hallmark.url.pathways <- paste0('https://raw.githubusercontent.com/SCCC-BBC',
                   '/PathwaySplice/development/inst/extdata',
                     '/h.all.v6.0.symbols.gmt.txt')
hup <- gmtGene2Cat(hallmark.url.pathways, genomeID='hg19')}

}
