% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goseq2.R
\name{goseq2}
\alias{goseq2}
\title{goseq2}
\usage{
goseq2(pwf, genome, id, gene2cat = NULL, test.cats = c("GO:CC", "GO:BP",
  "GO:MF"), method = "Wallenius", repcnt = 2000,
  use_genes_without_cat = FALSE)
}
\arguments{
\item{use_genes_without_cat}{}
}
\description{
goseq2
}
\examples{

sink("test_out.txt")
test.goseq2<-goseq2(Re.Go.adjusted.by.number.junction.2[[2]],"mm10","ensGene")

test.goseq3<-goseq2(Re.Go.adjusted.by.number.junction.2[[2]],"mm10","ensGene")

test.goseq4<-goseq2(Re.Go.adjusted.by.number.junction.2[[2]],"mm10","ensGene")

test.goseq5<-goseq2(Re.Go.adjusted.by.number.junction.2[[2]],"mm10","ensGene")

sink()

}

