% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goseq2.R
\name{goseq2}
\alias{goseq2}
\title{goseq2}
\usage{
goseq2(pwf, genome, id, gene.model, gene2cat = NULL, test.cats = c("GO:CC",
  "GO:BP", "GO:MF"), method = "Wallenius", repcnt = 2000,
  use_genes_without_cat = FALSE)
}
\arguments{
\item{pwf:}{probability weight function}

\item{genome:}{genome you use}

\item{id:}{which gene id}

\item{gene2cat:}{a list with gene as name and category as value}

\item{test.cats:}{the category you choose}

\item{method:}{which method for calculting overpresentation p-value}

\item{repcnt:}{the number of replications}

\item{use_genes_without_cat:}{whether using genes without category or not}
}
\description{
Modifying goseq to generate GO term with the listed DEgene
}
\examples{
data(gene.model)
gene_model<-gene.model

GO.wall.DE_interest.geneGL=goseq2(Gene.based.DE.feature.based.DE$pwfGeneGL,"mm10","ensGene",gene.model=gene_model)

GO.wall.DE_interes.geneFT=goseq2(Gene.based.DE.feature.based.DE$pwfGeneFeature,"mm10","ensGene",gene.model=gene_model)

GO.wall.DE_interest.FtFT=goseq2(Gene.based.DE.feature.based.DE$pwfFeatureFeature,"mm10","ensGene",gene.model=gene_model)

}

