% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotermAnalysis2GetAllGOTerms_AdjustedByNumOfJunctionWithinOneGene.R
\name{GotermAnalysis2GetAllGOTerms_AdjustedByNumOfJunctionWithinOneGene}
\alias{GotermAnalysis2GetAllGOTerms_AdjustedByNumOfJunctionWithinOneGene}
\title{This function is to subset data set based on feature}
\usage{

  GotermAnalysis2GetAllGOTerms_AdjustedByNumOfJunctionWithinOneGene(re.gene.based,
  ad = "GL", sub_feature = NULL, threshold, file_prefix)
}
\arguments{
\item{file_prefix}{}
}
\description{
This function is to subset data set based on feature
}
\examples{


Re.Go.adjusted.by.number.junction.2<-GotermAnalysis2GetAllGOTerms_AdjustedByNumOfJunctionWithinOneGene(
re.PJ.gene.based,ad="J","J",0.05,"Splice_junction_based")

data.pwf2.SJs<-plotPWF2(Re.Go.adjusted.by.number.junction.2[[2]],binsize=30,xlab = "Number of SJs(<binsize> gene bins)")


}

