% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotermAnalysis2GetAllGOTerms_AdjustedByNumOfJunctionWithinOneGene.R
\name{GotermAnalysis2GetAllGOTerms_AdjustedByNumOfJunctionWithinOneGene}
\alias{GotermAnalysis2GetAllGOTerms_AdjustedByNumOfJunctionWithinOneGene}
\title{Title}
\usage{
GotermAnalysis2GetAllGOTerms_AdjustedByNumOfJunctionWithinOneGene(Data4Goterm,
  ad = "GL", sub_feature = NULL, threshold, file_prefix)
}
\arguments{
\item{file_prefix}{}
}
\description{
Title
}

