% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotermAnalysisUseFeatureDefineDE.R
\name{GotermAnalysisUseFeatureDefineDE}
\alias{GotermAnalysisUseFeatureDefineDE}
\title{GotermAnalysisUseFeatureDefineDE
Descrption: This function computes statistical significance of GO terms, given a list of significant genes. These significant genes can be directly from differntial splicing analysis software such as rMat, DEXSeq or JunctionSeq. 
Based on genewise results to perform Go term analysis using DGEs defined by different criterion}
\usage{
GotermAnalysisUseFeatureDefineDE(re.gene.based, ad = "GL",
  sub_feature = NULL, DE_define = c("Feature", "GeneWise", "rMAT",
  "FeatureGeneWise", "FeaturerMAT", "GeneWiserMAT", "FeatureGeneWiseRMAT"),
  gene_model, Output_file_dir)
}
\arguments{
\item{gene_model}{}
}
\description{
GotermAnalysisUseFeatureDefineDE
Descrption: This function computes statistical significance of GO terms, given a list of significant genes. These significant genes can be directly from differntial splicing analysis software such as rMat, DEXSeq or JunctionSeq. 
Based on genewise results to perform Go term analysis using DGEs defined by different criterion
}
\examples{


Re.Go.adjusted.by.number.junction.2<-GotermAnalysis2GetAllGOTerms_AdjustedByNumOfJunctionWithinOneGene(
re.PJ.gene.based,ad="J","J",0.05,"Splice_junction_based")

data.pwf2.SJs<-plotPWF2(Re.Go.adjusted.by.number.junction.2[[2]],binsize=30,xlab = "Number of SJs(<binsize> gene bins)")

Re.Go.adjusted.by.exon.SJ<-GotermAnalysis2GetAllGOTerms_AdjustedByNumOfJunctionWithinOneGene(
re.PJ.gene.based,ad="exon_SJ",sub_feature=NULL,0.05,file_prefix="Exon_Splice_junction_based.xls",gene_model=gene.model)

}

