% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GotermAnalysisUseFeatureDefineDE.R
\name{GotermAnalysisUseFeatureDefineDE}
\alias{GotermAnalysisUseFeatureDefineDE}
\title{GotermAnalysisUseFeatureDefineDE}
\usage{
GotermAnalysisUseFeatureDefineDE(re.gene.based, ad = "GL",
  sub_feature = NULL, DE_define = c("Feature", "GeneWise", "rMAT",
  "FeatureGeneWise", "FeaturerMAT", "GeneWiserMAT", "FeatureGeneWiseRMAT"),
  gene_model, Output_file_dir)
}
\arguments{
\item{gene_model}{}
}
\description{
GotermAnalysisUseFeatureDefineDE
}
\examples{


Re.Go.adjusted.by.number.junction.2<-GotermAnalysis2GetAllGOTerms_AdjustedByNumOfJunctionWithinOneGene(
re.PJ.gene.based,ad="J","J",0.05,"Splice_junction_based")

data.pwf2.SJs<-plotPWF2(Re.Go.adjusted.by.number.junction.2[[2]],binsize=30,xlab = "Number of SJs(<binsize> gene bins)")

Re.Go.adjusted.by.exon.SJ<-GotermAnalysis2GetAllGOTerms_AdjustedByNumOfJunctionWithinOneGene(
re.PJ.gene.based,ad="exon_SJ",sub_feature=NULL,0.05,file_prefix="Exon_Splice_junction_based.xls",gene_model=gene.model)

}

