% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSA.read.gmt.2.R
\name{GSA.read.gmt.2}
\alias{GSA.read.gmt.2}
\title{GSA.read.gmt.2}
\usage{
GSA.read.gmt.2(filename)
}
\arguments{
\item{gmt}{filename}
}
\description{
GSA.read.gmt.2
}
\examples{

re.gsa<-GSA.read.gmt.2("/media/H_driver/Annotation/MsigDB/c2.cp.Mouse.v5.1.symbols.gmt")

re.ggsa.2<-GSA.read.gmt("/media/H_driver/Annotation/MsigDB/c2.cp.Mouse.v5.1.symbols.gmt")
}

