% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JS.perGeneQValue.R
\name{JS.perGeneQValue}
\alias{JS.perGeneQValue}
\title{JS.perGeneQValue}
\usage{
JS.perGeneQValue(pvals, wTest, geneID, method = JS.perGeneQValueExact)
}
\arguments{
\item{method}{}
}
\description{
JS.perGeneQValue
}
\examples{

data.perGeneQ.use.large.FC.featutre<-JS.perGeneQValue(Re.PJ.selected$pvalue,
Re.PJ.selected$testable,Re.PJ.selected$geneID,method = JS.perGeneQValueExact)

length(unique(as.character(Re.PJ.selected$geneID)))

head(Re.PJ.selected)


data.perGeneQ.use.large.FC.featutre.2<-list_to_df(data.perGeneQ.use.large.FC.featutre)
colnames(data.perGeneQ.use.large.FC.featutre.2)<-c("geneID","geneWisePadj_FC")

Re.PJ.selected.2<-merge(Re.PJ.selected,data.perGeneQ.use.large.FC.featutre.2,by="geneID")


head(Re.PJ.selected.2)

re.PJ.gene.based.selected<-merge(pData(re.PJ.gene.based),data.perGeneQ.use.large.FC.featutre.2,by="geneID")
}

