% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{lrTestBias}
\alias{lrTestBias}
\title{lrTestBias}
\usage{
lrTestBias(genewise.table, boxplot.width = 0.1)
}
\arguments{
\item{genewise.table}{A dataframe with genewise p-value for each gene, returned from \code{makeGeneTable()}}

\item{boxplot.width}{width of boxplot}

\item{loc.x}{x coordinate for position of logistic regression p-value in figure}

\item{loc.y}{y coordinate for position of logistic regression p-value in figure}

\item{y.lim}{The largest number of gene features on the y axis in boxplot}
}
\description{
This function tests presence of selection bias using logistic regression, and produces a boxplot 
that compares distributions of bias factors (e.g. number of exons) for significant genes and non-significant genes.
}
\details{
To determine presentce of selection bias, we fit the logistic regression model 
\code{Pr(a gene is significant) ~ number of features within the gene}. 
Here features refer to exon bins or splicing junction bins, depending on 
how genewise pvalues were obtained
}
\examples{
gene.based.table <- makeGeneTable(featureBasedData)
lrTestBias(gene.based.table)


}
