% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lcb.R
\name{lrtestbias}
\alias{lrtestbias}
\title{lrtestbias}
\usage{
lrtestbias(jscs_genewise_object, genewise.pvalue = "geneWisePadj",
  sig.threshold = 0.05, type = c("exon", "splicing"), loc.x = 2,
  loc.y = 70, y_lim = 80, boxplot_width)
}
\arguments{
\item{jscs_genewise_object}{Gene based table returned
from ReformatData and MakeGeneWiseTable_JunctionSeq}

\item{genewise.pvalue}{Whether you use smallest p value or geneWisePadj}

\item{sig.threshold}{Threshold to define differential gene list}

\item{type}{Whether you are interested in exon or splicing junction}

\item{loc.x}{Indicates x coordinate of p value from logistic regression}

\item{loc.y}{Indicates y coordinate of p value from logistic regression}

\item{y_lim}{Defining the largest number of exons in y axis in boxplot}

\item{boxplot_width}{Parameter for boxplot width}
}
\value{
Results from logistic regression
}
\description{
Logistic regression to check bias
}
\examples{
res <- mds.11.sample[which(as.numeric(mds.11.sample$numExons)<=50),]
res1 <- lrtestbias(res,loc.x=2,loc.y=70,y_lim=80,boxplot_width=0.3)
}
