% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lcb.R
\name{LRtestBias}
\alias{LRtestBias}
\title{LRtestBias}
\usage{
LRtestBias(jscs_genewise_object, genewise.pvalue = "geneWisePadj",
  sig.threshold = 0.05, type = c("exon", "splicing"), p.x, p.y, y_lim,
  boxplot_width)
}
\arguments{
\item{jscs_genewise_object}{gene based table returned
from ReformatData and MakeGeneWiseTable_JunctionSeq}

\item{genewise.pvalue}{whether you use smallest p value or geneWisePadj}

\item{sig.threshold}{threshold to define differential gene list}

\item{type}{whether you are interested in exon or splicing junction}

\item{p.x}{x coordinate for p value label}

\item{p.y}{y coordinate for p value label}

\item{boxplot_width}{parameter for boxplot width}
}
\value{
return results from logistic regression
}
\description{
Logistic regression to check bias
}
\examples{

data(mds)
re<-LRtestBias(mds,boxplot_width=0.3)

data(TinyData)
re<-LRtestBias(tiny.data,p.x=2,p.y=150,boxplot_width=0.3)

re<-LRtestBias(mds2,p.x=2,p.y=600,boxplot_width=0.3)

head(mds2,100)

range(as.numeric(mds2$numExons))
hist(as.numeric(mds2$numExons))

boxplot(as.numeric(mds2$numExons))

mds3<-mds2[which(as.numeric(mds2$numExons)<=50),]

hist(as.numeric(mds3$numExons))

dim(mds3)
dim(mds2)

mds<-mds3

save(mds,file="./data/mds.RData")

dim(mds)
hist(as.numeric(mds$numExons))
re<-LRtestBias(mds3,p.x=2,p.y=70,y_lim=80,boxplot_width=0.3)

}

