% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{makeGeneTable}
\alias{makeGeneTable}
\title{makeGeneTable}
\usage{
makeGeneTable(feature.table, sig.threshold = 0.05, stat = "pvalue")
}
\arguments{
\item{feature.table}{An \code{featureBasedData} object.}

\item{sig.threshold}{Significance threshold used to determine whether the gene is significant 
or not}

\item{stat}{The statistic used to select significant genes. Options are "pvalue" or "fdr"}
}
\value{
Returns a genewised table with several variables (columns) 
\item{geneID}{Gene identifiers in ensembl gene IDs}
\item{geneWisePvalue}{each gene is represented by the smallest p-value among its features}
\item{sig.gene}{a gene is significant (1) or not (0)} 
\item{mostSigDeFeature}{the most significant gene feature}
\item{numFeature}{number of gene features within the gene}
}
\description{
This function obtains genewise p-values,
by representing each gene with the smallest p-value among its features, 
and then determines genes status as significant or not.
}
\examples{
data(featureBasedData)
gene.based.table <- makeGeneTable(featureBasedData)
}
