% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{makeGeneWiseTable}
\alias{makeGeneWiseTable}
\title{makeGeneWiseTable}
\usage{
makeGeneWiseTable(jscs, gene.list = NULL, FDR.threshold = 0.05,
  verbose = TRUE)
}
\arguments{
\item{jscs}{An \code{JunctionSeqCountSet} object returned from function \code{getResultsFromJunctionSeq},
with restuls generated by testing differential usage of gene features (i.e. exon counting bins and/or splice junction counting bins)}

\item{gene.list}{A character vector of gene identifiers to be included in the returned table. 
If NULL, the genes included in the table will be unique genes in \code{jscs}}

\item{FDR.threshold}{Significance threshold to classify features (exons and/or splicing junction counting bins) as differentially used or not}

\item{verbose}{Whether to print process log or not}

\item{baseMean}{mean of the counts across samples in each feature}
}
\value{
A dataframe with the following arguments for each gene
}
\description{
This function generates a dataframe that summarizes (for each gene) differential usage of features within the gene.
}
\examples{

#First test differential usage for exons or splicing junctions
dir.name <- system.file('extdata',package = 'PathwaySplice')
sample.file <- 'Sample_info.txt'
count.file <- 'Counts.10.genes.txt'
gff.file <- 'flat.chr22.10.genes.gff'
res <- getResultsFromJunctionSeq(dir.name, sample.file, 
                                 count.file, gff.file, 
                                 method.dispFinal='max')
                                 
# Convert the results of differential usage analysis into a gene table
res1 <- makeGeneWiseTable(res)



}
