% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{makeGeneWiseTable}
\alias{makeGeneWiseTable}
\title{makeGeneWiseTable}
\usage{
makeGeneWiseTable(jscs, gene.list = NULL, FDR.threshold = 0.05,
  verbose = TRUE)
}
\arguments{
\item{jscs}{An \code{JunctionSeqCountSet} object returned from function \code{getResultsFromJunctionSeq},
with restuls generated by testing differential usage of gene features (i.e. exon counting bins and/or splice junction counting bins)}

\item{gene.list}{A character vector of gene identifiers to be included in the returned table. 
If NULL, the genes included in the table will be unique genes in \code{jscs}}

\item{FDR.threshold}{Significance threshold to classify features (exons and/or splicing junction counting bins) as differentially used or not}

\item{verbose}{Whether to print process log or not}
}
\value{
A dataframe with the following arguments for each gene 
       from GetResultsFromJunctionSeq with the followings as column names:
  \itemize{
  \item geneID: Gene ID
  \item chr: Chromosome information
  \item start: Starting position
  \item end: End position
  \item strand: Strand information
  \item baseMean: mean of the counts across samples in each feature
  \item geneWisePadj: perGeneQvalue from DEXSeq package 
  \item mostSigID: The feature that have smallest p-value within gene  
  \item mostSigPadjust: p-value for the most significant feature 
  \item numExons: Number of exons within gene
  \item numKnown: Number of splicing junction within gene
  \item numNovel: Number of novel splicing junction within gene
  \item exonsSig: Number of significant exons
  \item knownSig: Number of significant splicing junction
  \item novelSig: Number of significant novel splicing junction
  \item numFeatures: A list that includes number of exons, number of 
  splicing junction and number of novel splicing junction within gene 
  \item numSig: A list that includes number of significant exons, number of 
  significant splicing junction and number of significant novel splicing 
  junction within gene 
}
}
\description{
This function generates a dataframe that summarizes (for each gene) differential usage of features within the gene.
}
\examples{

#First test differential usage for exons or splicing junctions
dir.name <- system.file('extdata',package = 'PathwaySplice')
sample.file <- 'Sample_info.txt'
count.file <- 'Counts.10.genes.txt'
gff.file <- 'flat.chr22.10.genes.gff'
res <- getResultsFromJunctionSeq(dir.name, sample.file, 
                                 count.file, gff.file, 
                                 method.dispFinal='max')
                                 
# Convert the results of differential usage analysis into a gene table
res1 <- makeGeneWiseTable(res)



}
