% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGeneWiseTable.R
\name{makeGeneWiseTable}
\alias{makeGeneWiseTable}
\title{makeGeneWiseTable}
\usage{
makeGeneWiseTable(jscs, gene.list, FDR.threshold = 0.05, verbose = TRUE,
  debug.mode = FALSE)
}
\arguments{
\item{debug.mode}{}
}
\description{
This function is to generate the gene-based results
}
\examples{

# Generate genewise table for output differential exons or splicing junctions results

re.example.gene.based<-makeGeneWiseTable(Re.example,gene.list=unique(as.character(fData(Re.example)$geneID)))
re.example.gene.based.testable.reformat<-ReformatData(re.example.gene.based)

}

