% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeGeneWiseTable_JunctionSeq.R, R/makeGeneWiseTable.R
\name{makeGeneWiseTable}
\alias{makeGeneWiseTable}
\title{makeGeneWiseTable}
\usage{
makeGeneWiseTable(jscs, gene.list, FDR.threshold = 0.05, verbose = TRUE)

makeGeneWiseTable(jscs, gene.list, FDR.threshold = 0.05, verbose = TRUE)
}
\arguments{
\item{debug.mode}{}

\item{jscs:}{output from JunctionSeq}

\item{gene.list:}{gene list}

\item{FDR.threshold:}{threshold to define whether exon or splicing junction is differential used}

\item{verbose:}{whether to print runing process}
}
\description{
This function is to generate gene-based results

This function is used to generate the gene-based results
}
\examples{

re.example.gene.based<-makeGeneWiseTable(Re.example,gene.list=unique(as.character(fData(Re.example)$geneID)))
re.example.gene.based.testable.reformat<-ReformatData(re.example.gene.based)

save(re.example.gene.based,re.example.gene.based.testable.reformat,file=paste0(getwd(),"example_Mut_WT_gene.RData"))

mds<-re.example.gene.based.testable.reformat


re.example.gene.based<-makeGeneWiseTable(Re.example,gene.list=unique(as.character(fData(Re.example)$geneID)))
re.example.gene.based.testable.reformat<-ReformatData(re.example.gene.based)

}

