% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGeneWiseTable.R
\name{makeGeneWiseTable}
\alias{makeGeneWiseTable}
\title{makeGeneWiseTable}
\usage{
makeGeneWiseTable(jscs, gene.list = NULL, FDR.threshold = 0.05,
  verbose = TRUE)
}
\arguments{
\item{jscs}{The object returned from \code{"GetResultsFromJunctionSeq"}}

\item{gene.list}{Gene list of the returned object from JunctionSeq or DEXSeq}

\item{FDR.threshold}{Threshold to define the differential usage of either exon
or splicing junction}

\item{verbose}{Whether to print runing process or not}
}
\value{
A gene based table converted from the object returned 
       from GetResultsFromJunctionSeq
}
\description{
This function is used to generate the gene-based results
}
\examples{

dir.name <- system.file("extdata", package="PathwaySplice")
sample.file <- "Sample_info.txt"
count.file <- "QC.spliceJunctionAndExonCounts.forJunctionSeq.txt"
gff.file <- "flat.chr22.gff"
#res1 <- GetResultsFromJunctionSeq(dir.name, sample.file, count.file, gff.file)
#res2 <- makeGeneWiseTable(res1)

}
