% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGeneWiseTable.R
\name{makeGeneWiseTable}
\alias{makeGeneWiseTable}
\title{makeGeneWiseTable}
\usage{
makeGeneWiseTable(jscs, gene.list, FDR.threshold = 0.05, verbose = TRUE)
}
\arguments{
\item{jscs}{output from JunctionSeq}

\item{gene.list}{gene list}

\item{FDR.threshold}{threshold to define whether exon or splicing junction is differential used}

\item{verbose}{whether to print runing process}
}
\value{
A gene based table
}
\description{
This function is used to generate the gene-based results
}
\examples{

dir.name=dirname(system.file("extdata","decoder.bySample.Mut_WT_example.txt", package = "PathwaySplice"))
dir.name=paste0(dir.name,"/")
file.sample="decoder.bySample.Mut_WT_example.txt"
file.gff="Homo_sapiens.GRCh38.84.processed.sorted.4.JunctionSeq.flat.chr4.100.gff"
file.count="/QC.spliceJunctionAndExonCounts.forJunctionSeq.txt"
Re.example<-GetResultsFromJunctionSeq(dir.name,file.sample,file.count,file.gff)
all.gene.list<-unique(as.character(fData(Re.example)$geneID))
choosed.gene.list<-all.gene.list
re.example.gene.based<-makeGeneWiseTable(Re.example,
gene.list=choosed.gene.list)


}

