% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGeneWiseTable.R
\name{makeGeneWiseTable}
\alias{makeGeneWiseTable}
\title{makeGeneWiseTable}
\usage{
makeGeneWiseTable(jscs, gene.list, FDR.threshold = 0.05, verbose = TRUE)
}
\arguments{
\item{jscs:}{output from JunctionSeq}

\item{gene.list:}{gene list}

\item{FDR.threshold:}{threshold to define whether exon or splicing junction is differential used}

\item{verbose:}{whether to print runing process}
}
\description{
This function is used to generate the gene-based results
}
\examples{

#re.example.gene.based<-makeGeneWiseTable(Re.example,gene.list=unique(as.character(fData(Re.example)$geneID)))


}

