% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGeneWiseTable.R
\name{makeGeneWiseTable}
\alias{makeGeneWiseTable}
\title{makeGeneWiseTable}
\usage{
makeGeneWiseTable(jscs, gene.list, FDR.threshold = 0.05, verbose = TRUE)
}
\arguments{
\item{jscs}{The object returned from \code{"GetResultsFromJunctionSeq"}}

\item{gene.list}{Gene list of the returned object from JunctionSeq or DEXSeq}

\item{FDR.threshold}{Threshold to define the differential usage of either exon
or splicing junction}

\item{verbose}{Whether to print runing process or not}
}
\value{
A gene based table converted from the object returned 
       from GetResultsFromJunctionSeq
}
\description{
This function is used to generate the gene-based results
}
\examples{
library(Biobase)
file.sample="Sample_info.txt"
dir.name=dirname(system.file("extdata","Sample_info.txt", package = "PathwaySplice"))
dir.name=paste0(dir.name,"/")
file.gff="flat.chr22.gff"
file.count="/QC.spliceJunctionAndExonCounts.forJunctionSeq.txt"

\donttest{Re.example<-GetResultsFromJunctionSeq(dir.name,file.sample,file.count,file.gff)

re.example.gene.based<-makeGeneWiseTable(Re.example,
gene.list=unique(as.character(fData(Re.example)$geneID)))}


}
