% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeGeneWiseTable_JunctionSeq.R
\name{makeGeneWiseTable}
\alias{makeGeneWiseTable}
\title{This function is to generate the gene-based results}
\usage{
makeGeneWiseTable(jscs, gene.list, FDR.threshold = 0.05, verbose = TRUE,
  debug.mode = FALSE)
}
\arguments{
\item{debug.mode}{}
}
\description{
This function is to generate the gene-based results
}

