% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGeneWiseTable.R
\name{makeGeneWiseTable}
\alias{makeGeneWiseTable}
\title{makeGeneWiseTable}
\usage{
makeGeneWiseTable(jscs, gene.list, FDR.threshold = 0.05, verbose = TRUE)
}
\arguments{
\item{jscs}{output from JunctionSeq}

\item{gene.list}{gene list}

\item{FDR.threshold}{threshold to define whether exon or splicing junction is differential used}

\item{verbose}{whether to print runing process}
}
\value{
A gene based table
}
\description{
This function is used to generate the gene-based results
}
\examples{

all.gene.list<-unique(as.character(fData(Re.example)$geneID))

#make a tiny example data set
choosed.gene.list<-sample(all.gene.list,500)
re.example.gene.based<-makeGeneWiseTable(Re.example,
gene.list=choosed.gene.list)
tiny.data<-re.example.gene.based
save(tiny.data,file="./data/TinyData.RData")

#make an example data using all genes

all.gene.data<-makeGeneWiseTable(Re.example,
gene.list=all.gene.list)
mds2<-all.gene.data
save(mds2,file="./data/mds2.RData")


}

