% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGeneWiseTable.R
\name{makegenewisetable}
\alias{makegenewisetable}
\title{makegenewisetable}
\usage{
makegenewisetable(jscs, gene.list = NULL, FDR.threshold = 0.05,
  verbose = TRUE)
}
\arguments{
\item{jscs}{The object returned from \code{'GetResultsFromJunctionSeq'}}

\item{gene.list}{Gene list of the returned object from JunctionSeq or DEXSeq}

\item{FDR.threshold}{Threshold to define the differential usage of either exon
or splicing junction}

\item{verbose}{Whether to print runing process or not}
}
\value{
A gene based table converted from the object returned 
       from GetResultsFromJunctionSeq
}
\description{
This function is used to generate the gene-based results
}
\examples{

#Get differential usage for exons or splicing junctions
dir.name <- system.file('extdata',package = 'PathwaySplice')
sample.file <- 'Sample_info.txt'
count.file <- 'Counts.10.genes.txt'
gff.file <- 'flat.chr22.10.genes.gff'
res <- getresultsfromjunctionseq(dir.name, sample.file, 
count.file, gff.file, method.dispFinal='max')
# Convert the results of differential usage analysis into gene based resutls
res1 <- makegenewisetable(res)
}
