% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{makeGeneWiseTable}
\alias{makeGeneWiseTable}
\title{makeGeneWiseTable}
\usage{
makeGeneWiseTable(jscs, gene.list = NULL, FDR.threshold = 0.05,
  verbose = TRUE)
}
\arguments{
\item{jscs}{The object returned from \code{'getResultsFromJunctionSeq'}}

\item{gene.list}{Gene list of the returned object from JunctionSeq or DEXSeq}

\item{FDR.threshold}{Threshold to define the differential usage of either exon
or splicing junction}

\item{verbose}{Whether to print runing process or not}
}
\value{
A gene based table converted from the object returned 
       from GetResultsFromJunctionSeq with the followings as column names:
  \itemize{
  \item geneID: Gene ID
  \item chr: Chromosome information
  \item start: Starting position
  \item end: End position
  \item strand: Strand information
  \item baseMean: BaseMean based on all samples
  \item geneWisePadj: perGeneQvalue from DEXSeq package 
  \item mostSigID: The feature that have smallest p-value within gene  
  \item mostSigPadjust: p-value for the most significant feature 
  \item numExons: Number of exons within gene
  \item numKnown: Number of splicing junction within gene
  \item numNovel: Number of novel splicing junction within gene
  \item exonsSig: Number of significant exons
  \item knownSig: Number of significant splicing junction
  \item novelSig: Number of significant novel splicing junction
  \item numFeatures: A list that includes number of exons, number of 
  splicing junction and number of novel splicing junction within gene 
  \item numSig: A list that includes number of significant exons, number of 
  significant splicing junction and number of significant novel splicing 
  junction within gene 
}
}
\description{
This function is used to generate the gene-based results
}
\examples{

#Get differential usage for exons or splicing junctions
dir.name <- system.file('extdata',package = 'PathwaySplice')
sample.file <- 'Sample_info.txt'
count.file <- 'Counts.10.genes.txt'
gff.file <- 'flat.chr22.10.genes.gff'
res <- getResultsFromJunctionSeq(dir.name, sample.file, 
                                 count.file, gff.file, 
                                 method.dispFinal='max')
# Convert the results of differential usage analysis into gene based results
res1 <- makeGeneWiseTable(res)
}
