% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{outKegg2Gmt}
\alias{outKegg2Gmt}
\title{outKegg2gmt}
\usage{
outKegg2Gmt(organism.id, out.gmt.file)
}
\arguments{
\item{organism.id}{an identifier for the organism being studied, for example, "hsa" for "Homo sapiens"}

\item{out.gmt.file}{name of the output .gmt file}
}
\value{
Returns a .gmt file for KEGG pathways
}
\description{
This function obtains a .gmt file for KEGG pathways.
}
\details{
The function calls the \code{get.kegg.genesets} function in \code{EnrichmentBrowser} R package 
and modifies the resulting output into a .gmt file.
}
\examples{
\dontrun{

data.dir <- system.file ("extdata", package = "PathwaySplice")
outKegg2Gmt ("hsa",file.path(data.dir,"kegg.gmt.txt"))
}

}
