% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPWF_modified.R, R/plotPWF2.R
\name{plotPWF2}
\alias{plotPWF2}
\title{Plot the probability Weighting Function}
\usage{
plotPWF2(pwf, binsize = "auto", pwf_col = 3, pwf_lwd = 2,
  xlab = "Biased Data in <binsize> gene bins.", ylab = "Proportion DE", ...)

plotPWF2(pwf, binsize = "auto", pwf_col = 3, pwf_lwd = 2,
xlab = "Biased Data in <binsize> gene bins.", ylab = "Proportion DE", ...)
}
\arguments{
\item{pwf}{probability weigth function}

\item{binsize}{the number of gene in each bin(gene set)}

\item{pwf_col}{the color for the fitted line of probability weigth function}

\item{pwf_lwd}{the font for the fitted line of probability weigth function}

\item{xlab}{the label for x-axis}

\item{ylab}{the label for y-axis}

\item{...}{}

\item{...}{}
}
\description{
Plot the probability Weighting Function
}

