% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{PostProcessGO}
\alias{PostProcessGO}
\title{PostProcessGO}
\usage{
PostProcessGO(n.go, adjusted, unadjuasted, venn.dir, boxplot.dir,
  In.ad.not.un.file, In.un.not.ad.file)
}
\arguments{
\item{n.go}{number of gene sets}

\item{adjusted}{adjusted result}

\item{unadjuasted}{unadjusted result}

\item{venn.dir}{path for outputing venn}

\item{boxplot.dir}{path for outputing boxplot}

\item{In.ad.not.un.file}{file name for outputing adjused but not in unadjusted when using the selected gene sets}

\item{In.un.not.ad.file}{file name for outputing unadjused but not in adjusted when using the selected gene sets}
}
\value{
null
}
\description{
PostProcessGO
}

