% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{postProcessGo}
\alias{postProcessGo}
\title{postProcessGo}
\usage{
postProcessGo(n.go, adjusted, unadjuasted, venn.dir, boxplot.dir,
  type.boxplot = c("All", "Only3"), In.ad.not.un.file, In.un.not.ad.file)
}
\arguments{
\item{n.go}{Number of gene sets}

\item{adjusted}{Adjusted result}

\item{unadjuasted}{Unadjusted result}

\item{venn.dir}{Path for outputing venn}

\item{boxplot.dir}{Path for outputing boxplot}

\item{type.boxplot}{Get boxplot for 5 categories or 3 categories
6 categories: 'All.adjusted','All.unadjusted',
              'Top25.adjusted','Top25.unadjusted',
              'In_ad_not_un','In_un_not_ad'
3 categories: 'All','Top25.adjusted','Top25.unadjusted}

\item{In.ad.not.un.file}{File name for outputing adjused but not
in unadjusted when using the selected gene sets}

\item{In.un.not.ad.file}{File name for outputing unadjused but not
in adjusted when using the selected gene sets}
}
\value{
null
}
\description{
postProcessGo
}
\examples{

dir.name <- system.file('extdata', package='PathwaySplice')
canonical.pathway.file <- '10.cp.gmt.txt'
res <- gmtGene2Cat(dir.name,canonical.pathway.file,'local',genomeID='hg19')
gene.based.table <- makeGeneTable(featureBasedData)

res1 <- runPathwaySplice(gene.based.table,genome='hg19',id='ensGene',gene2cat=res,method='Wallenius')

res2 <- runPathwaySplice(gene.based.table,genome='hg19',id='ensGene',gene2cat=res,method='Hypergeometric')

dir.name <- tempdir()
output.dir <- file.path(dir.name,'OutputPostAnalysis')

output.file.name.1 <- 'In_ad_not_un.xls'
output.file.name.2 <- 'In_un_not_ad.xls'
res3 <- postProcessGo(4,res1,res2,output.dir,output.dir,
                      type.boxplot='Only3',
                      output.file.name.1,output.file.name.2)
}
