% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_function_4_PathwaySplice.R
\name{reformatpath}
\alias{reformatpath}
\title{reformatpath}
\usage{
reformatpath(dir.name)
}
\arguments{
\item{dir.name}{Directory name to be converted}
}
\value{
A converted directory
}
\description{
reformatpath
}
\examples{
dir.name <- '/media/H_driver/2016/Yang/MACS/MACS/'
converted.dir.name <- PathwaySplice:::reformatpath(dir.name)

}
