% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{runpathwaysplice}
\alias{runpathwaysplice}
\title{runpathwaysplice}
\usage{
runpathwaysplice(re.gene.based, adjust = "GL", sub_feature = NULL,
  threshold, genomeID, geneID, method, gene2cat = NULL)
}
\arguments{
\item{re.gene.based}{Gene based results}

\item{adjust}{Bias factor to be adjusted}

\item{sub_feature}{Feature to be checked}

\item{threshold}{Threshold to be used for adjustment}

\item{genomeID}{Genome to be used(hg19 or mm10)}

\item{geneID}{GeneID to be used(entrezgene or ensembl_gene_id)}

\item{method}{Method to be used for calculating overrepresented p value
of gene sets(Options include Wallenius,Sampling, and Hypergeometric)}

\item{gene2cat}{Get sets defined by users}
}
\value{
A list that has gene set enrichment analysis results
}
\description{
Perform analysis using Run_pathwaysplice,user use gene-based
table as an input, and select genes based on subfeature, then choose
to using which bias factor to perform adjutsment
}
\examples{
res <- runpathwaysplice(tiny.data,adjust='exon_SJ',sub_feature='E',
0.05,genomeID='hg19',geneID='ensGene', method='Wallenius')

}
