% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{runPathwaySplice}
\alias{runPathwaySplice}
\title{runPathwaySplice}
\usage{
runPathwaySplice(res, genome, id, gene2cat = NULL, test.cats = c("GO:CC",
  "GO:BP", "GO:MF"), go.size.cut = c(lower.size = 0, upper.size = NULL),
  method = "Wallenius", repcnt = 2000, use.genes.without.cat = FALSE)
}
\arguments{
\item{res}{Gene based table}

\item{genome}{Genome to be used(hg19 or mm10)}

\item{id}{GeneID to be used(entrezgene or ensembl_gene_id)}

\item{gene2cat}{Get sets defined by users}

\item{test.cats}{Gene set if users does not define their gene set}

\item{go.size.cut}{Size of gene sets to be defined}

\item{method}{Method to be used for calculating overrepresented p value
of gene sets(Options include Wallenius,Sampling, and Hypergeometric)}

\item{repcnt}{Number of sampling if user use sampling method to calculate gene set enrichment p value}

\item{use.genes.without.cat}{Whether to include gene without mapping to get set for calculate gene set enrichment p value. if set FALSE, use the genes that have mapped gene sets only as background; if set TRUE, use all genes as background}
}
\value{
A list that has gene set enrichment analysis results
}
\description{
This function uses gene-based
table converted from makeGeneTable as an input, and select gene sets of certain type and size  
to perform gene set enrichment analysis adjusted by number of features in gene based table
}
\examples{
res <- makeGeneTable(featureBasedData)
res <- runPathwaySplice(res,genome='hg19',id='ensGene',
                         test.cats=c('GO:BP'),
                         go.size.cut=c(5,30),
                         method='Wallenius')

 
}
