% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{runPathwaySplice}
\alias{runPathwaySplice}
\title{runPathwaySplice}
\usage{
runPathwaySplice(genewise.table, genome, id, gene2cat = NULL,
  test.cats = c("GO:CC", "GO:BP", "GO:MF"), go.size.cut = c(-Inf, Inf),
  method = "Wallenius", repcnt = 2000, use.genes.without.cat = FALSE)
}
\arguments{
\item{genewise.table}{data frame returned from function \code{makeGeneTable}}

\item{genome}{Genome to be used, options are "hg19" or "mm10"}

\item{id}{GeneID, options are "entrezgene" or "ensembl_gene_id"}

\item{gene2cat}{Get sets defined by users}

\item{test.cats}{Gene sets to be tested if \code{gene2cat} is not defined}

\item{go.size.cut}{Size limit of the gene sets to be tested}

\item{method}{the method used to calculate pathway enrichment p value. 
Options are "Wallenius", "Sampling", and "Hypergeometric"}

\item{repcnt}{Number of random samples to be calculated when "Sampling" is used
ignored unless \code{method="Sampling"}}

\item{use.genes.without.cat}{Whether genes not mapped to any category tested are included in analysis.
If set to FALSE, genes not mapped to any tested categories are ignored in analysis.}
}
\value{
A list that has gene set enrichment analysis results
}
\description{
This function identifies pathways that are enriched with signficant genes, while accounting for 
different number of exons and/or splicing junctions associated with each gene
}
\examples{
res <- makeGeneTable(featureBasedData)
res <- runPathwaySplice(res,genome='hg19',id='ensGene',
                         test.cats=c('GO:BP'),
                         go.size.cut=c(5,30),
                         method='Wallenius')

 
}
