% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{runPathwaySplice}
\alias{runPathwaySplice}
\title{runPathwaySplice}
\usage{
runPathwaySplice(genewise.table, genome, id, gene2cat = NULL,
  test.cats = c("GO:CC", "GO:BP", "GO:MF"), go.size.limit = c(10, 200),
  method = "Wallenius", repcnt = 2000, use.genes.without.cat = FALSE,
  binsize = "auto")
}
\arguments{
\item{genewise.table}{data frame returned from function \code{makeGeneTable}}

\item{genome}{Genome to be used, options are 'hg19' or 'mm10'}

\item{id}{GeneID, options are 'entrezgene' or 'ensembl_gene_id'}

\item{gene2cat}{Get sets defined by users, can be obtained for example from \code{gmtGene2Cat} function}

\item{test.cats}{Default gene sets to be tested if \code{gene2cat} is not defined}

\item{go.size.limit}{Size limit of the gene sets to be tested}

\item{method}{the method used to calculate pathway enrichment p value. 
Options are 'Wallenius', 'Sampling', and 'Hypergeometric'}

\item{repcnt}{Number of random samples to be calculated when 'Sampling' is used
ignored unless \code{method='Sampling'}}

\item{use.genes.without.cat}{Whether genes not mapped to any category tested are included in analysis.
If set to FALSE, genes not mapped to any tested categories are ignored in analysis.}

\item{binsize}{The number of genes in each gene bin in the bias plot}
}
\value{
runPathwaySplice returns a tibble(data frame) with 11 columns as the following:
\item{category}{Name of the gene set (e.g. a pathway or a gene category)} 
\item{over_represented_pvalue}{P-vaue for the associated category being over represented among significant genes} 
\item{under_represented_pvalue}{P-vaue for the associated category being under represented among significant genes} 
\item{numDEInCat}{The number of significant genes in the category} 
\item{numInCat}{The total number of genes in the category}                                          
\item{description}{The GO term if any of the categories is a GO term} 
\item{ontology}{The column for the GO term's ontology if any of the categories is a GO term}
\item{DEgene_ID}{The column for the Ensembl gene ID of differentially genes in the category}
\item{DEgene_symbol}{The column for the gene symbol of differentially genes in the category}
\item{Ave_value_all_gene}{The column for the average numFeature value of total genes in the category}
}
\description{
This function identifies pathways that are enriched with signficant genes, while accounting for 
different number of gene features (e.g. exons) associated with each gene
}
\details{
This function implements the methodology described in Young et al. (2011) to adjust for 
         different number of gene features (i.e. counting bins, see Fig 1 in Anders et al. 2012) associated with each gene. In the bias plot, the genes are grouped 
         by \code{numFeature} in \code{genewise.table} into gene bins, 
         the proportions of signficant genes are then plotted against the gene bins.
}
\examples{
gene.based.table <- makeGeneTable(featureBasedData)
res <- runPathwaySplice(gene.based.table,genome='hg19',id='ensGene',
                         test.cats=c('GO:BP'),
                         go.size.limit=c(5,30),
                         method='Wallenius',binsize=2)

 
}
\references{
Young MD, Wakefield MJ, Smyth GK, Oshlack A (2011) \emph{Gene ontology analysis for RNA-seq: 
accounting for selection bias}. Genome Biology 11:R14

Anders S, Reyes A, Huber W (2012) \emph{Dececting differential usage of exons from RNA-seq data.} 
Genome Research 22(10): 2008-2017
}
