% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{runPathwaySplice2}
\alias{runPathwaySplice2}
\title{Title}
\usage{
runPathwaySplice2(res2, genome, id, gene2cat = NULL, test.cats = c("GO:CC",
  "GO:BP", "GO:MF"), go.size.cut = c(lower.size = 0, upper.size = NULL),
  method = "Wallenius", repcnt = 2000, use.genes.without.cat = FALSE)
}
\arguments{
\item{res2}{}

\item{gene2cat}{}

\item{method}{}

\item{genomeID}{}

\item{geneID}{}
}
\description{
Title
}
\examples{

res4 <- PathwaySplice:::runPathwaySplice2(res2,genome='hg19',id='ensGene',test.cats=c("GO:CC"),go.size.cut=c(5,30),method='Wallenius')

 
}
