% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_pathwaysplice.R
\name{runPathwaySplice2}
\alias{runPathwaySplice2}
\title{Title}
\usage{
runPathwaySplice2(res2, genomeID, geneID, method, gene2cat = NULL)
}
\arguments{
\item{res2}{}

\item{genomeID}{}

\item{geneID}{}

\item{method}{}

\item{gene2cat}{}
}
\description{
Title
}
\examples{

res4 <- runPathwaySplice2(res2,genomeID='hg19',geneID='ensGene', 
                        method='Wallenius')

}
