% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationSJ2DE.R, R/SimulationSJ2DE2.R
\name{SimulationSJ2DE}
\alias{SimulationSJ2DE}
\title{Title}
\usage{
SimulationSJ2DE(min_num_splicing, max_num_splicing, num_gene)

SimulationSJ2DE(min_num_splicing, max_num_splicing, num_gene)
}
\arguments{
\item{num_gene}{}

\item{num_gene}{}
}
\description{
Title
}
\examples{
re.random.DE2SJ<-SimulationSJ2DE(20,200,30)
}

