% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testEnrichMap.R
\name{testpathwaysplice}
\alias{testpathwaysplice}
\title{testPathwaySplice}
\usage{
testpathwaysplice(gene_based_table, output.file.dir)
}
\arguments{
\item{gene_based_table}{A gene based table converted from 
DEXSeq or JunctionSeq resutls}

\item{output.file.dir}{Directory for output}
}
\value{
None
}
\description{
Peform one-step analysis for check bias, adjusted gene set 
enrichment anlysis and build network
}
\examples{

output.file.dir <- "~/OutputTest"
testpathwaysplice(tiny.data,output.file.dir = output.file.dir)


}
