% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testEnrichMap.R
\name{testPathwaySplice}
\alias{testPathwaySplice}
\title{testPathwaySplice}
\usage{
testPathwaySplice(gene_based_table, which_gene_model)
}
\arguments{
\item{gene_based_table}{A gene based table converted from DEXSeq or JunctionSeq resutls}

\item{which_gene_model}{Gene annotation file}
}
\description{
Peform one-step analysis for check bias, adjusted gene set enrichment anlysis and build network
}
\examples{
data(mds)
data(hg19) 

testPathwaySplice(mds,hg19.gene.model)


}

