% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testEnrichMap.R
\name{testpathwaysplice}
\alias{testpathwaysplice}
\title{testPathwaySplice}
\usage{
testpathwaysplice(gene_based_table)
}
\arguments{
\item{gene_based_table}{A gene based table converted from 
DEXSeq or JunctionSeq resutls}
}
\value{
None
}
\description{
Peform one-step analysis for check bias, adjusted gene set 
enrichment anlysis and build network
}
\examples{
testpathwaysplice(tiny.data)


}
