% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doc4Data.R
\docType{data}
\name{tiny.data}
\alias{tiny.data}
\title{tiny.data}
\format{A data frame with 252 rows and 17 variables}
\usage{
data(TinyData)
}
\value{
A data set with 252 rows and 17 variables
}
\description{
A small gene-based data set generated by sampling 500 features from the count file including all features.
}
\details{
We used Hartley's QoRTs program to get counts of exons and/or splicing

junctions for each feature. The following is the example table

'ENSG00000135226+ENSG00000215110' for the sample SRR1660311:

ENSG00000135226+ENSG00000215110:A000 9

ENSG00000135226+ENSG00000215110:E001 0

ENSG00000135226+ENSG00000215110:E002 0

ENSG00000135226+ENSG00000215110:E003 0

...

ENSG00000135226+ENSG00000215110:E005 2

ENSG00000135226+ENSG00000215110:E006 0

ENSG00000135226+ENSG00000215110:E007 7

...

ENSG00000135226+ENSG00000215110:J016 0

...

ESG00000135226+ENSG00000215110:J028  0

We applied Hartley's JunctionSeq on counts to calculate Differential

Usage(DU) of exons and/or splicing junction, then converted

DU results to gene based table. Since we include the features that could

be "aggregate genes" which consist of multiple genes, and certain genes are included

in some "aggregate genes" mutiple times, so we obtain the information of

252 genes from 500 features after we covert DU resutls to gene based

table.

This gene based table includes the following information:

\itemize{

\item geneID: Gene ID
  \item chr: Chromosome information
  \item start: Starting position
  \item end: End position
  \item strand: Strand information
  \item baseMean: BaseMean based on all samples
  \item geneWisePadj: perGeneQvalue from DEXSeq package 
  \item mostSigID: The feature that have smallest p-value within gene  
  \item mostSigPadjust: p-value for the most significant feature 
  \item numExons: Number of exons within gene
  \item numKnown: Number of splicing junction within gene
  \item numNovel: Number of novel splicing junction within gene
  \item exonsSig: Number of significant exons
  \item knownSig: Number of significant splicing junction
  \item novelSig: Number of significant novel splicing junction
  \item numFeatures: A list that includes number of exons, number of 
  splicing junction and number of novel splicing junction within gene 
  \item numSig: A list that includes number of significant exons, number of 
  significant splicing junction and number of significant novel splicing 
  junction within gene 
}

In this gene based table, for 'ENSG00000135226+ENSG00000215110',

we give row names as 'ENSG00000135226.ENSG00000135226+ENSG00000215110'

and 'ENSG00000215110.ENSG00000135226+ENSG00000215110' to demonstrate

'ENSG00000135226' and 'ENSG00000215110' are two genes from

'ENSG00000135226+ENSG00000215110'.
}
\keyword{datasets}
